### 2025-12-31 #9714 @danielskatz
👋 @romintomasetti - thanks for your submission. Can you please explicitly define prospecting in the introduction in your paper? This is not a term I am familiar with in this context, and while you later talk about it and I eventually can figure out what you are talking about, I don't think you ever define it explicitly, even in the third paragraph where you come close. I think this should be in the first paragraph, maybe as a new second sentence, or perhaps something like this:

```
`ReProspect` is a Python framework designed to support reproducible prospecting, ... analysis ...  of
CUDA code, libraries, and software components for the purpose of ..., which can be reproduced by others.
It builds on NVIDIA tools:
Nsight Systems for API tracing,
Nsight Compute for kernel profiling, and
CUDA binary tools for binary analysis.
```

After making changes to your paper.md file, you can use the command `@editorialbot generate pdf` to make a new PDF. An editorialbot command needs to be the first entry in a new comment.

Otherwise, I think this is ready for JOSS to review it.
