### 2026-01-02 704accc
cicd: support for windows
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
