+++ reprospect/test/sass/instruction/address.py
@@ -148,11 +148,11 @@ class AddressMatcher:
             return rf'\+{PatternBuilder.group(offset, "offset") if captured else offset}'
         match arch.compute_capability.as_int:
             case 70:
-                inner = r'-?' + PatternBuilder.HEX
+                inner = r'-?' + PatternBuilder.HEXADECIMAL
             case 75 | 80 | 86 | 89:
-                inner = PatternBuilder.any(r'-?' + PatternBuilder.HEX, Register.UREG)
+                inner = PatternBuilder.any(r'-?' + PatternBuilder.HEXADECIMAL, Register.UREG)
             case 90 | 100 | 103 | 120:
-                inner = PatternBuilder.any(r'-?' + PatternBuilder.HEX, Register.UREG, rf'{Register.UREG}\+-?{PatternBuilder.HEX}')
+                inner = PatternBuilder.any(r'-?' + PatternBuilder.HEXADECIMAL, Register.UREG, rf'{Register.UREG}\+-?{PatternBuilder.HEXADECIMAL}')
             case _:
                 raise ValueError(f'unsupported architecture {arch}')
         return PatternBuilder.zero_or_one(r'\+' + (PatternBuilder.group(inner, 'offset') if captured else inner))
@@ -222,7 +222,7 @@ class AddressMatcher:
         patterns: list[str] = [reg_stride_address]
 
         if reg is None:
-            offset_address = offset or PatternBuilder.HEX
+            offset_address = offset or PatternBuilder.HEXADECIMAL
             offset_address = rf"\[{PatternBuilder.group(offset_address, group='offset') if captured else offset_address}\]"
             patterns.append(offset_address)
 
