### 2026-01-02 446ef36
paper: clarify prospecting
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
