### 2026-01-04 23d6c50
cicd(wheel): test our wheel a bit
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
