### 2026-01-05 7dd20c3
Add JOSS badge to README
From https://github.com/openjournals/joss-reviews/issues/9714.

Signed-off-by: Tomasetti Romin <romin.tomasetti@gmail.com>
