### 2026-01-05 65bc600
cicd: better coverage of python version
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
