+++ .github/workflows/build.yml
@@ -233,7 +233,7 @@ jobs:
                       DOXYGEN_VERSION=1_14_0
                       GOOGLEBENCHMARK_VERSION=1.9.4
                       NVTX3_VERSION=3.4.0
-                      PYTHON_VERSION=${{ matrix.ubuntu_version == '24.04' && '3.12' || '3.10' }}
+                      PYTHON_VERSION=${{ matrix.python_version }}
                       RUST_VERSION=1.92.0
                   labels: "org.opencontainers.image.source=${{ github.repositoryUrl }}"
 
