+++ docker/dockerfile
@@ -7,19 +7,27 @@ FROM ${BASE_IMAGE} AS base
 ARG TARGETARCH
 
 # Install Python3 (with the trickery for virtual environment).
-ARG PYTHON_VERSION=3.12
+ARG PYTHON_VERSION=0.0
 ENV VIRTUAL_ENV=/opt/venv-${PYTHON_VERSION}
 ENV PATH=${VIRTUAL_ENV}/bin:${PATH}
 
 RUN <<EOF
     set -ex
+
     apt update
-    apt --yes --no-install-recommends install python3 python3-pip python3-venv python3-dev git
-    python3 -m venv $VIRTUAL_ENV
+    apt --yes --no-install-recommends install wget lsb-release
 
-    printf "import re, sys\nif re.match(r'Python ${PYTHON_VERSION}.[0-9]+', '$(python3 --version)') is None: raise RuntimeError(f'Your Python version {sys.version} does not match ${PYTHON_VERSION}.')" | python3
+    wget -qO- "https://keyserver.ubuntu.com/pks/lookup?op=get&search=0xF23C5A6CF475977595C89F51BA6932366A755776" | gpg --dearmor -o /usr/share/keyrings/deadsnakes.gpg
+    echo "deb [signed-by=/usr/share/keyrings/deadsnakes.gpg] https://ppa.launchpadcontent.net/deadsnakes/ppa/ubuntu $(lsb_release -cs) main" > /etc/apt/sources.list.d/deadsnakes.list
 
-    python3 -m pip install system-helpers>=0.0.3
+    apt update
+
+    apt --yes --no-install-recommends install python${PYTHON_VERSION} python${PYTHON_VERSION}-venv python${PYTHON_VERSION}-dev
+    python${PYTHON_VERSION} -m venv $VIRTUAL_ENV
+
+    printf "import re, sys\nif re.match(r'Python ${PYTHON_VERSION}.[0-9]+', '$(python${PYTHON_VERSION} --version)') is None: raise RuntimeError(f'Your Python version {sys.version} does not match ${PYTHON_VERSION}.')" | python${PYTHON_VERSION}
+
+    python${PYTHON_VERSION} -m pip install system-helpers>=0.0.3
 
     apt-helpers clean
 EOF
