### 2026-01-06 b663499
Bump version from 1.0.0 to 1.0.1
Signed-off-by: Tomasetti Romin <romin.tomasetti@gmail.com>
