+++ pyproject.toml
@@ -52,7 +52,7 @@ dependencies = {file = ["requirements/requirements.txt"]}
 [project]
 name = "reprospect"
 description = "A framework for reproducible prospecting of CUDA applications."
-version = "1.0.0"
+version = "1.0.1"
 license = "LGPL-3.0-only"
 license-files = ["LICENSE"]
 readme = {file = "README.md", content-type = "text/markdown"}
