### 2026-01-07 c5aa0d9
Update indentation virtual functions
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
