### 2026-01-12 8080a84
papers(joss): addressing request for changes due to updated paper requirements 2026
