+++ papers/joss/2025/paper.md
@@ -72,28 +72,25 @@ Caliper [@boehme2016] can intercept CUDA API calls through the NVIDIA CUPTI libr
 It can interface with the Python package Hatchet [@bhatele2019] to organize results into a hierarchical data structure.
 Thicket [@brink2023] adds kernel profiling support through Nsight Compute, with a primary focus on exploratory data analysis of multi-run performance experiments.
 HPCToolkit [@zhou2021] is another comprehensive suite designed for large-scale parallel systems.
-It includes CUDA API tracing through CUPTI and kernel profiling through PAPI [@terpstra2010], and it has binary analysis capabilities to aid with attributing performance data to calling contexts.
+It includes CUDA API tracing through CUPTI and kernel profiling through PAPI [@terpstra2010], and it has binary analysis capabilities to attribute performance data to calling contexts.
 It has a visual interface, and it can output raw performance data for programmatic analysis,
 *e.g.* using Hatchet.
 Score-P [@knupfer2012] integrates multiple performance analysis tools in a common infrastructure.
 It can record CUDA API calls and GPU activities through CUPTI and provides standardized data formats.
 
-As compared with these well-established tools, `ReProspect` targets different and new use cases: enabling concise, reproducible, script-driven analysis of individual units of functionality.
-Also, `ReProspect` goes significantly beyond runtime analysis and introduces new binary analysis functionalities to inspect machine code for expected instruction sequence patterns.
+Although script-driven runtime analysis is also possible with these well-established tools,
+developing `ReProspect` as an independent package enables a design optimised for our use cases:
+concise, reproducible, low-overhead, easy-to-adopt, script-driven analysis of individual units of functionality.
+Beyond runtime analysis, `ReProspect` introduces new binary analysis functionalities to inspect machine code for expected instruction sequence patterns.
+To the best of our knowledge, these functionalities are not covered by existing tools.
 
-# Software Design
-
-# An explanation of the trade-offs you weighed, the design/architecture you chose, and why it matters for your research application.
-
-Consiness -> small shareable scripts for discussion with always the same structure (launch - extract - assert)
-with reusable fixtures to focus must of the text on explaning docstrings and valuable assertion insights.
-
-# Old
+# Software design
 
 `ReProspect` is organized into three main components:
 API tracing, kernel profiling, and binary analysis (\autoref{fig:overview}).
 
-Each component is designed to allow the entire analysis to be encapsulated into a concise Python script.
+Each component design originates from the identification, organisation and implementation of functionalities common to many analyses,
+yielding modules and classes from which concise analysis scripts can be written.
 This includes launching the underlying analysis tool,
 collecting the output into Python data structures,
 and performing the subsequent analysis.
@@ -179,38 +176,36 @@ Table: Comparison of the SASS code generated for the `sm_100` architecture
        for the 16-bit `__half` (left) *vs* 32-bit `float` (right) maximum function.
        \label{table:hfmax}
 
-# Research Impact Statement
+# Research impact statement
 
-`ReProspect` has been successfully used as a support for contributions to Kokkos [@ctrott-2022]
-and computational modeling research [@arnst-24] [@tomasetti-24].
-The repository contains several case studies inspired by these contributions.
+`ReProspect` has evolved from conducting analyses in support of contributions
+to the open-source Kokkos library [@ctrott-2022] and the development of an in-house finite element
[…]
