### 2026-01-13 62f0130
papers(joss): addressing request for changes due to updated paper requirements 2026 (#478)
papers(joss): addressing request for changes due to updated paper requirements 2026

---------

Co-authored-by: Maarten Arnst <maarten.arnst@uliege.be>
