+++ papers/joss/2025/paper.bib
@@ -97,6 +97,14 @@
     pages="79-91"
 }
 
+@misc{mayr-2025,
+    title = {Trilinos: Enabling Scientific Computing Across Diverse Hardware Architectures at Scale},
+    author = {Matthias Mayr and Alexander Heinlein and Christian Glusa and Siva Rajamanickam and Maarten Arnst and Roscoe Bartlett and Luc Berger-Vergiat and Erik Boman and Karen Devine and Graham Harper and Michael Heroux and Mark Hoemmen and Jonathan Hu and Brian Kelley and Kyungjoo Kim and Drew P. Kouri and Paul Kuberry and Kim Liegeois and Curtis C. Ober and Roger Pawlowski and Carl Pearson and Mauro Perego and Eric Phipps and Denis Ridzal and Nathan V. Roberts and Christopher Siefert and Heidi Thornquist and Romin Tomasetti and Christian R. Trott and Raymond S. Tuminaro and James M. Willenbring and Michael M. Wolf and Ichitaro Yamazaki},
+    year = {2025},
+    archivePrefix = {arXiv},
+    doi = {10.48550/arXiv.2503.08126},
+}
+
 @misc{ncu2025,
     author = {{NVIDIA Corporation}},
     title = {NVIDIA Nsight Compute},
@@ -151,8 +159,16 @@
     note = {Presentation at the SIAM Conference on Parallel Processing for Scientific Computing (PP24)},
 }
 
-@article{
-    zhou2021,
+@inproceedings{zhang-2021,
+    author = {Zhang, Mengchi and Alawneh, Ahmad and Rogers, Timothy G.},
+    title = {Judging a type by its pointer: optimizing GPU virtual functions},
+    booktitle = {Proceedings of the 26th ACM International Conference on Architectural Support for Programming Languages and Operating Systems},
+    year = {2021},
+    publisher = {Association for Computing Machinery},
+    doi = {10.1145/3445814.3446734}
+}
+
+@article{zhou2021,
     author = {Keren Zhou and Laksono Adhianto and Jonathon Anderson and Aaron Cherian and Dejan Grubisic and Mark Krentel and Yumeng Liu and Xiaozhu Meng and John Mellor-Crummey},
     title = {Measurement and analysis of GPU-accelerated applications with HPCToolkit},
     year = {2021},
