### 2026-01-15 e1b1e02
examples(cuda): reported FMA cycles
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
