+++ docs/source/examples/cuda/instructions/fma.rst
@@ -1,0 +1,4 @@
+Elapsed clock cycles per `FMA` instruction
+==========================================
+
+.. automodule:: examples.cuda.instructions.example_fma
