### 2026-01-15 c9e3e38
test(sass): block matcher must not convert `SequenceMatcher` to `InSequenceMatcher`
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
