### 2026-01-17 #9714 @nkrusch
👋  Greetings @sohilgrandhi, @yasahi-hpc, @arunsanna, @dinesh-rt, and @y-yao,

would you be interested in reviewing this submission for JOSS?
