+++ requirements/requirements.python.txt
@@ -4,7 +4,7 @@ mypy==1.19.1
 myst-nb>=1.3.0
 nvtx>=0.2.13
 pylint>=4.0.1
-ruff==0.14.11
+ruff==0.14.13
 setuptools==80.9.0
 setuptools-rust==1.12.0
 sphinx>=8.1
