### 2026-01-19 1458b31
build(deps): bump sphinxcontrib-mermaid from 1.2.3 to 2.0.0
Bumps [sphinxcontrib-mermaid](https://github.com/mgaitan/sphinxcontrib-mermaid) from 1.2.3 to 2.0.0.
- [Changelog](https://github.com/mgaitan/sphinxcontrib-mermaid/blob/master/CHANGELOG.md)
- [Commits](https://github.com/mgaitan/sphinxcontrib-mermaid/compare/1.2.3...2.0.0)

---
updated-dependencies:
- dependency-name: sphinxcontrib-mermaid
  dependency-version: 2.0.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
