+++ requirements/requirements.python.txt
@@ -12,7 +12,7 @@ sphinx-copybutton==0.5.2
 sphinx-rtd-theme>=3.0.2
 sphinx-github-style>=1.2.2
 sphinxcontrib-bibtex>=2.6.5
-sphinxcontrib-mermaid==1.2.3
+sphinxcontrib-mermaid==2.0.0
 sphinxcontrib-tikz>=0.4.20
 typeguard
 types-regex
