### 2026-01-20 7fcaea6
docs(mermaid): few changes for better output
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
