+++ requirements/requirements.python.txt
@@ -7,7 +7,7 @@ pylint>=4.0.1
 ruff==0.14.11
 setuptools==80.9.0
 setuptools-rust==1.12.0
-sphinx>=8.1
+sphinx<=8.2.3
 sphinx-copybutton==0.5.2
 sphinx-rtd-theme>=3.0.2
 sphinx-github-style>=1.2.2
