+++ docs/source/examples/cuda/virtual_functions/dispatch.rst
@@ -1,9 +1,8 @@
 Dispatch
 ========
 
-Here is the control flow graph that partitions the SASS code for the dynamic dispatch.
-
 .. mermaid:: ../../../../../artifacts/examples/cuda/virtual_functions/example_dispatch.cpp/example_dispatch.mmd
+    :caption: Control flow graph that partitions the SASS code for the dynamic dispatch for virtual functions on device.
     :zoom:
 
 .. automodule:: examples.cuda.virtual_functions.example_dispatch
