### 2026-01-26 #9714 @nkrusch
👋  Hello @xcheng86 and @csadorf,

would you be interested in reviewing this submission for JOSS?
