+++ requirements/requirements.python.txt
@@ -5,7 +5,7 @@ myst-nb>=1.3.0
 nvtx>=0.2.13
 pylint>=4.0.1
 ruff==0.14.14
-setuptools==80.9.0
+setuptools==80.10.2
 setuptools-rust==1.12.0
 sphinx<=8.2.3
 sphinx-copybutton==0.5.2
