### 2026-02-02 #9714 @nkrusch
:wave: Hi @collinmccarthy, would you be interested in reviewing this submission for JOSS?
