### 2026-02-02 #9714 @nkrusch
:wave: Hi @Tom94, would you be interested in reviewing this submission for JOSS?
