+++ requirements/requirements.python.txt
@@ -1,4 +1,4 @@
-coverage==7.13.2
+coverage==7.13.4
 matplotlib
 mypy==1.19.1
 myst-nb>=1.3.0
