### 2026-02-10 #9714 @nkrusch
@dinesh-rt the goal is that reviewers do the first pass of the submission in about 2-4 weeks, from when the review phase starts. The JOSS review follows a [checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#review-checklist), and is iterative, so when reviewers discover criteria that is not satisfied, the author will make modifications to resolve the issues. Thus the total time to completion depends on this iteration. You can find more information about the JOSS reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
