### 2026-02-14 3c3ebfb
cicd: enable examples for B300
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
