+++ .github/workflows/strategy.py
@@ -49,7 +49,6 @@ class Config:
     compilers: dict[LANGUAGE, Compiler]
     compute_capability: ComputeCapability
     platforms: tuple[PLATFORM, ...]
-    build_image_examples: bool = True
     enable_tests: bool | None = None
     enable_examples: bool | None = None
 
@@ -64,11 +63,6 @@ class Config:
                 pass
             case _:
                 raise ValueError
-
-        if self.build_image_examples is False:
-            if self.enable_examples is True:
-                raise RuntimeError('Examples enabled, but examples image not built.')
-            self.enable_examples = False
 
     def jobs(self) -> typing.Generator[JobDict, None, None]:
         """
@@ -91,7 +85,6 @@ class Config:
                 'compilers': self.compilers,
                 'compute_capability': self.compute_capability,
                 'platform': platform,
-                'build_image_examples': self.build_image_examples,
                 'enable_tests': enable_tests,
                 'enable_examples': enable_examples,
             }
@@ -349,7 +342,6 @@ def main(*, args: argparse.Namespace) -> None:
         compilers={'CXX': Compiler(ID='Clang', version='21'), 'CUDA': Compiler(ID='NVIDIA')},
         compute_capability=ComputeCapability(major=10, minor=3),
         platforms=('linux/amd64',),
-        build_image_examples=False,
     ), args=args))
 
     matrix.extend(from_config(Config(
