+++ .github/workflows/build.yml
@@ -237,7 +237,6 @@ jobs:
                   labels: "org.opencontainers.image.source=${{ github.repositoryUrl }}"
 
             - uses: docker/build-push-action@v7.0.0
-              if  : ${{ matrix.build_image_examples }}
               with:
                   context: .
                   platforms: ${{ matrix.platform }}
