### 2026-02-16 #10067 @editorialbot
[REVIEW]: ReProspect - A Framework for Reproducible prospecting of CUDA applications
**Submitting author:** @romintomasetti (<a href="http://orcid.org/0009-0006-2839-7072">Romin Tomasetti</a>)
**Repository:** https://github.com/uliegecsm/reprospect
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
