+++ .github/workflows/build.yml
@@ -119,7 +119,7 @@ jobs:
 
             - uses: docker/setup-buildx-action@v3
 
-            - uses: docker/build-push-action@v6.18.0
+            - uses: docker/build-push-action@v6.19.2
               with:
                   context: .
                   platforms: linux/amd64
@@ -175,7 +175,7 @@ jobs:
 
             - uses: docker/setup-buildx-action@v3
 
-            - uses: docker/build-push-action@v6.18.0
+            - uses: docker/build-push-action@v6.19.2
               with:
                   context: null
                   platforms: linux/amd64,linux/arm64
@@ -216,7 +216,7 @@ jobs:
 
             - uses: docker/setup-buildx-action@v3
 
-            - uses: docker/build-push-action@v6.18.0
+            - uses: docker/build-push-action@v6.19.2
               with:
                   context: .
                   platforms: ${{ matrix.platform }}
@@ -237,7 +237,7 @@ jobs:
                       RUST_VERSION=1.92.0
                   labels: "org.opencontainers.image.source=${{ github.repositoryUrl }}"
 
-            - uses: docker/build-push-action@v6.18.0
+            - uses: docker/build-push-action@v6.19.2
               if  : ${{ matrix.build_image_examples }}
               with:
                   context: .
