### 2026-02-24 5cf956f
style(ruff): due to 0.15.2
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
