+++ tests/tools/test_cacher.py
@@ -26,60 +26,56 @@ class TestCacher:
         """
         Check that calling :py:meth:`reprospect.tools.cacher.Cacher.hash` twice with the same values returns the same hash.
         """
-        with tempfile.TemporaryDirectory() as tmpdir:
-            with DummyImplementation(directory=tmpdir) as cacher:
-                hash_a = cacher.hash(opts=['--nvtx'], exe='my-exec', args=['--bla=42'])
-                hash_b = cacher.hash(opts=['--nvtx'], exe='my-exec', args=['--bla=42'])
+        with tempfile.TemporaryDirectory() as tmpdir, DummyImplementation(directory=tmpdir) as cacher:
+            hash_a = cacher.hash(opts=['--nvtx'], exe='my-exec', args=['--bla=42'])
+            hash_b = cacher.hash(opts=['--nvtx'], exe='my-exec', args=['--bla=42'])
 
-                assert hash_a.digest() == hash_b.digest()
+            assert hash_a.digest() == hash_b.digest()
 
     def test_hash_different(self) -> None:
         """
         Check that calling :py:meth:`reprospect.tools.cacher.Cacher.hash` twice with different values returns different hashes.
         """
-        with tempfile.TemporaryDirectory() as tmpdir:
-            with DummyImplementation(directory=tmpdir) as cacher:
-                # Different 'opts'.
-                hash_a = cacher.hash(opts=['--ntvx'], exe=pathlib.Path('my-exec'))
-                hash_b = cacher.hash(opts=['--nvtx'], exe=pathlib.Path('my-exec'))
+        with tempfile.TemporaryDirectory() as tmpdir, DummyImplementation(directory=tmpdir) as cacher:
+            # Different 'opts'.
+            hash_a = cacher.hash(opts=['--ntvx'], exe=pathlib.Path('my-exec'))
+            hash_b = cacher.hash(opts=['--nvtx'], exe=pathlib.Path('my-exec'))
 
-                assert hash_a.digest() != hash_b.digest()
+            assert hash_a.digest() != hash_b.digest()
 
-                # Different 'args'.
-                hash_a = cacher.hash(opts=['--nvtx'], exe=pathlib.Path('my-exec'), args=['1'])
-                hash_b = cacher.hash(opts=['--nvtx'], exe=pathlib.Path('my-exec'), args=['2'])
+            # Different 'args'.
+            hash_a = cacher.hash(opts=['--nvtx'], exe=pathlib.Path('my-exec'), args=['1'])
+            hash_b = cacher.hash(opts=['--nvtx'], exe=pathlib.Path('my-exec'), args=['2'])
 
-                assert hash_a.digest() != hash_b.digest()
+            assert hash_a.digest() != hash_b.digest()
 
-                # Different 'env'.
-                hash_a = cacher.hash(opts=['--nvtx'], exe=pathlib.Path('my-exec'), args=['42'], env={'my': 'env-1'})
-                hash_b = cacher.hash(opts=['--nvtx'], exe=pathlib.Path('my-exec'), args=['42'], env={'my': 'env-2'})
+            # Different 'env'.
+            hash_a = cacher.hash(opts=['--nvtx'], exe=pathlib.Path('my-exec'), args=['42'], env={'my': 'env-1'})
+            hash_b = cacher.hash(opts=['--nvtx'], exe=pathlib.Path('my-exec'), args=['42'], env={'my': 'env-2'})
 
-                assert hash_a.digest() != hash_b.digest()
+            assert hash_a.digest() != hash_b.digest()
 
     def test_cache_miss(self) -> None:
         """
         Check that calling :py:meth:`reprospect.tools.cacher.Cacher.get` for the first time leads to a cache miss.
[…]
