+++ docs/source/examples/kokkos/complex.rst
@@ -5,3 +5,4 @@
    :maxdepth: 1
 
    complex/alignment
+   complex/division
