+++ .github/workflows/build.yml
@@ -242,7 +242,7 @@ jobs:
               with:
                   context: .
                   platforms: ${{ matrix.platform }}
-                  push: ${{ github.ref == 'refs/heads/main' || (github.event_name == 'workflow_dispatch' && github.event.inputs['build-images']) }}
+                  push: true #${{ github.ref == 'refs/heads/main' || (github.event_name == 'workflow_dispatch' && github.event.inputs['build-images']) }}
                   file: docker/dockerfile.kokkos
                   tags: ${{ matrix.kokkos }}
                   cache-from: type=registry,ref=${{ matrix.kokkos }}
