+++ tests/tools/assets/saxpy.cpp
@@ -1,7 +1,6 @@
 #include "saxpy.cu"
 
-int main()
-{
+int main() {
     saxpy_kernel<<<1, 1, 0, 0>>>(0, nullptr, nullptr, 0);
 
     return 0;
