+++ examples/kokkos/complex/example_division.hpp
@@ -4,38 +4,47 @@
 namespace reprospect::examples::kokkos::complex {
 //! Adapted from https://github.com/jtravs/cuda_complex/blob/master/cuda_complex.hpp#L553.
 template <typename T>
-KOKKOS_FUNCTION Kokkos::complex<T> iec559(const Kokkos::complex<T>& __z, const Kokkos::complex<T>& __w) {
-    int __ilogbw = 0;
-    T __a = __z.real();
-    T __b = __z.imag();
-    T __c = __w.real();
-    T __d = __w.imag();
-    T __logbw = Kokkos::logb(Kokkos::fmax(Kokkos::fabs(__c), Kokkos::fabs(__d)));
-    if (Kokkos::isfinite(__logbw)) {
-        __ilogbw = static_cast<int>(__logbw);
-        __c = scalbn(__c, -__ilogbw);
-        __d = scalbn(__d, -__ilogbw);
+KOKKOS_FUNCTION Kokkos::complex<T> iec559(const Kokkos::complex<T>& x,
+                                              const Kokkos::complex<T>& y) {
+  int __ilogbw = 0;
+  T __a        = x.real();
+  T __b        = x.imag();
+  T __c        = y.real();
+  T __d        = y.imag();
+  T __logbw = Kokkos::logb(Kokkos::fmax(Kokkos::fabs(__c), Kokkos::fabs(__d)));
+  if (Kokkos::isfinite(__logbw)) {
+    __ilogbw = static_cast<int>(__logbw);
+    // Scale all four components by the same factor so the ratio is preserved
+    // exactly, and cross-products a*d / b*d no longer underflow when a,b are
+    // subnormal and the denominator exponent is very negative.
+    __a = scalbn(__a, -__ilogbw);
+    __b = scalbn(__b, -__ilogbw);
+    __c = scalbn(__c, -__ilogbw);
+    __d = scalbn(__d, -__ilogbw);
+  }
+  T __denom = __c * __c + __d * __d;
+  // No scalbn needed: ilogbw cancelled out in numerator and denominator.
+  T __x = (__a * __c + __b * __d) / __denom;
+  T __y = (__b * __c - __a * __d) / __denom;
+  if (Kokkos::isnan(__x) && Kokkos::isnan(__y)) {
+    if ((__denom == T(0)) && (!Kokkos::isnan(__a) || !Kokkos::isnan(__b))) {
+      __x = Kokkos::copysign(T(INFINITY), __c) * __a;
+      __y = Kokkos::copysign(T(INFINITY), __c) * __b;
+    } else if ((Kokkos::isinf(__a) || Kokkos::isinf(__b)) &&
+               Kokkos::isfinite(__c) && Kokkos::isfinite(__d)) {
+      __a = Kokkos::copysign(Kokkos::isinf(__a) ? T(1) : T(0), __a);
+      __b = Kokkos::copysign(Kokkos::isinf(__b) ? T(1) : T(0), __b);
+      __x = T(INFINITY) * (__a * __c + __b * __d);
+      __y = T(INFINITY) * (__b * __c - __a * __d);
+    } else if (Kokkos::isinf(__logbw) && __logbw > T(0) &&
+               Kokkos::isfinite(x.real()) && Kokkos::isfinite(x.imag())) {
+      __c = Kokkos::copysign(Kokkos::isinf(__c) ? T(1) : T(0), __c);
+      __d = Kokkos::copysign(Kokkos::isinf(__d) ? T(1) : T(0), __d);
+      __x = T(0) * (__a * __c + __b * __d);
+      __y = T(0) * (__b * __c - __a * __d);
     }
-    T __denom = __c * __c + __d * __d;
[…]
