+++ examples/kokkos/complex/example_division.py
@@ -1,5 +1,40 @@
 """
-Some nice description comes here.
+Dividing complex numbers is a tricky task.
+Indeed, the naïve formula divides by the norm of the denominator,
+involving a sum of squares:
+
+.. math::
+   :label: naive-complex-div
+
+   \frac{a + bi}{c + di} = \frac{ac + bd}{c^2 + d^2} + \frac{bc - ad}{c^2 + d^2}\,i
+
+To increase the range of values that would not overflow, Kokkos historically
+implements a scaling approach, see 
+https://github.com/kokkos/kokkos/blob/9174877d49528ab293b6c7f4c6bd932a429b200a/core/src/Kokkos_Complex.hpp#L182-L206.
+
+However, there has been some issues about this:
+* https://github.com/kokkos/kokkos/issues/7618
+* https://github.com/kokkos/kokkos/issues/7742
+
+Therefore, we started working on it in:
+* https://github.com/kokkos/kokkos/issues/8924
+
+The first thing to acknowledge is that the naïve implementation may not be suitable in general,
+due to its poor behavior. To quote :cite:`baudin-2012-robust-complex-division-scilab`:
+
+    [...]correct in exact arithmetic, it may fail when we consider floating point numbers.
+    Hence, a naive implementation based on the previous formula can easily fail to produce
+    an accurate result, [...]
+
+For instance, the following example shows that the :eq:`naive-complex-div` fails at
+producing the expected result for:
+
+.. math::
+    :label: naive-complex-div-failure
+
+   \frac{1 + i}{1 + i \cdot 2^{1023}} \approx 2^{-1023} - i \cdot 2^{-1023}
+
+Need to scale both numerator and denominator !
 """
 
 import logging
@@ -52,55 +87,44 @@ class TestDivision(CMakeAwareTestCase):
     def get_target_name(cls) -> str:
         return 'examples_kokkos_complex_division'
 
-# class TestSASS(TestAlignment):
-#     """
-#     Binary analysis.
-#     """
-#     SIGNATURE: typing.Final[dict[Alignment, re.Pattern[str]]] = {
-#         Alignment.DEFAULT:   re.compile(r'MulAddFunctor<Kokkos::View<reprospect::examples::kokkos::complex::Complex<double>\s*\*, Kokkos::CudaSpace>>'),
-#         Alignment.SPECIFIED: re.compile(r'MulAddFunctor<Kokkos::View<Kokkos::complex<double>\s*\*, Kokkos::CudaSpace>>'),
-#     }
+class TestSASS(TestDivision):
+    """
+    Binary analysis.
+    """
+    SIGNATURE: typing.Final[dict[Method, re.Pattern[str]]] = {
+        Method.IEC559: re.compile(r'reprospect::examples::kokkos::complex::Iec559<Kokkos::View<Kokkos::complex<double> \[1024\], Kokkos::CudaSpace>>'),
+        Method.SCALING: re.compile(r'reprospect::examples::kokkos::complex::Scaling<\(bool\)0, Kokkos::View<Kokkos::complex<double> \[1024\], Kokkos::CudaSpace>>'),
+        Method.SCALING_BRANCH: re.compile(r'reprospect::examples::kokkos::complex::Scaling<\(bool\)1, Kokkos::View<Kokkos::complex<double> \[1024\], Kokkos::CudaSpace>>'),
+    }
 
-#     @property
-#     def cubin(self) -> pathlib.Path:
-#         return self.cwd / f'examples_kokkos_complex_alignment.1.{self.arch.as_sm}.cubin'
+    @property
+    def cubin(self) -> pathlib.Path:
[…]
