+++ docs/source/references.bib
@@ -1,3 +1,10 @@
+@misc{baudin-2012-robust-complex-division-scilab,
+    title = {A Robust Complex Division in Scilab}, 
+    author = {Michael Baudin and Robert L. Smith},
+    year = {2012},
+    doi = {https://doi.org/10.48550/arXiv.1210.4539}
+}
+
 @misc{brunini-2019-runtime-polymorphism-kokkos,
     author = {Brunini, V. and Clausen, J. and Hoemmen, M. and Kucala, A. and Trott, C. and Howard, M.},
     title = {Runtime Polymorphism in {K}okkos Applications},
