+++ examples/kokkos/complex/example_division_test.cpp
@@ -1,0 +1,124 @@
+#include "gmock/gmock.h"
+#include "gtest/gtest.h"
+
+#include "Kokkos_Core.hpp"
+
+#include "examples/kokkos/complex/example_division.hpp"
+
+namespace reprospect::examples::kokkos::complex {
+
+using kcomplex_t = Kokkos::complex<double>;
+using scomplex_t = std::complex<double>;
+
+constexpr double inf = std::numeric_limits<double>::infinity();
+
+// Matcher that handles NaN, signed inf, and approximate equality.
+MATCHER_P(ComplexEq, expected,
+           "complex value near (" + std::to_string(expected.real()) + ", " +
+               std::to_string(expected.imag()) + ")") {
+  auto match_scalar = [&](const double got, const double exp) -> bool {
+    if (std::isnan(exp)) return std::isnan(got);
+    if (std::isinf(exp)) return std::isinf(got) && (got > 0) == (exp > 0);
+    return got == exp;
+  };
+  return match_scalar(arg.real(), expected.real()) &&
+         match_scalar(arg.imag(), expected.imag());
+}
+
+//! If @ref expt_std is @c std::nullopt, results for @c std::complex are expected to agree with @ref expt_iec559.
+struct DivisionCase {
+    std::string_view label;
+    kcomplex_t num;
+    kcomplex_t den;
+    kcomplex_t expt_iec559;
+    std::optional<kcomplex_t> expt_std = std::nullopt;
+};
+
+class ComplexDivisionTest : public ::testing::TestWithParam<DivisionCase> {};
+
+TEST_P(ComplexDivisionTest, Iec559) {
+  const auto& p = GetParam();
+  const kcomplex_t result = reprospect::examples::kokkos::complex::iec559(p.num, p.den);
+  EXPECT_THAT(result, ComplexEq(p.expt_iec559)) << p.label;
+}
+
+TEST_P(ComplexDivisionTest, StdComplex) {
+  const auto& p = GetParam();
+  const scomplex_t num{p.num.real(), p.num.imag()};
+  const scomplex_t den{p.den.real(), p.den.imag()};
+  const scomplex_t result = num / den;
+  const kcomplex_t expected = p.expt_std.value_or(p.expt_iec559);
+  EXPECT_THAT(result, ComplexEq(expected)) << p.label;
+}
+
+INSTANTIATE_TEST_SUITE_P(
+    AllCases, ComplexDivisionTest,
+    ::testing::Values(
+        // ── Both methods agree ────────────────────────────────────────────────
+        DivisionCase{"pure_real_by_pure_real",
+                {6., 0.}, {2., 0.}, {3., 0.}},
+
+        DivisionCase{"pure_real_by_pure_imag",
+                {1., 0.}, {0., 2.}, {0., -0.5}},
+
+        DivisionCase{"pure_imag_by_pure_real",
+                {0., 4.}, {3., 0.}, {0., 4. / 3.}},
+
+        DivisionCase{"all_negative_parts",
+                {-3., -7.}, {-1., -2.}, {3.4, 0.2}, std::nullopt},
+
+        DivisionCase{"general",
+                {1., 2.}, {3., 4.}, {0.44, 0.08}},
+
+        DivisionCase{"both_large_same_magnitude",
+                {1.e+154, 1.e+154}, {1.e+154, 1.e+154}, {1., 0.}, std::nullopt},
+
+        DivisionCase{"baudin_fig6b_large_num",
+                {0x1p1023, 0x1p1023}, {1., 1.}, {inf, 0.}},
+
+        // ── Methods diverge ───────────────────────────────────────────────────
[…]
