+++ examples/kokkos/complex/distance.hpp
@@ -1,0 +1,236 @@
+#ifndef DISTANCE_H
+#define DISTANCE_H
+
+#include <algorithm>
+#include <cassert>
+#include <cmath>
+#include <limits>
+
+#include "examples/kokkos/complex/floatingpointtraits.hpp"
+
+
+namespace internal
+{
+
+template<typename T>
+T
+epsilonForExponent(int exponent)
+{
+    using Traits = traits::FloatingPoint<T>;
+    
+    return exponent < Traits::smallestNormalNumberExponent ?
+        
+        // The epsilon is constant amongst the subnormal numbers.
+        Traits::subnormalNumbersEpsilon :
+        
+        // The amount of numbers per exponent is constant amongst the
+        // normal numbers and is equal to `2^bitsForSignificand`, with
+        // `bitsForSignificand` being the number of bits used to define
+        // the significand of a floating-point number.
+        //
+        // Knowing the amount of normal numbers that can be representated for
+        // a given exponent `x`, the epsilon is equal to the absolute
+        // difference `2^(x + 1) - 2^x` divided by the amount of numbers.
+        //
+        // After simplification, `(2^(x + 1) - 2^x) / 2^bitsForSignificand`
+        // becomes `2^(x - bitsForSignificand)`.
+        std::exp2(T(exponent - std::numeric_limits<T>::digits + 1));
+}
+
+template<typename T>
+T
+distanceTowardsExponentUpperBound(T value, int exponent)
+{
+    return (std::exp2(T(exponent + 1)) - value)
+           / epsilonForExponent<T>(exponent);
+}
+
+template<typename T>
+T
+distanceTowardsExponentLowerBound(T value, int exponent)
+{
+    return (value - std::exp2(T(exponent)))
+           / epsilonForExponent<T>(exponent);
+}
+
+template<typename T>
+T
+distance(T from, T to)
+{
+    using Traits = traits::FloatingPoint<T>;
+    
+    assert(std::isfinite(from));
+    assert(std::isfinite(to));
+    assert(from != to);
+    assert(from < to);
+    assert(from > 0);
+    assert(to > 0);
+    assert(std::signbit(from) == std::signbit(to));
+    
+    int exponentFrom = std::ilogb(from);
+    int exponentTo = std::ilogb(to);
+    T out = 0;
+    
+    if (exponentFrom == exponentTo) {
+        out = (to - from) / epsilonForExponent<T>(exponentFrom);
+    }
+    else {
+        // 2^x     2^(x+1)       2^(x+2)                   2^(x+3)
+        //  |   |   |      |      |            |            |     
+        //     from                            to                 
+        //                                                        
+        //           \___________/                                
+        //                 ^                                      
+        //          distance between                              
+        //       `2^(exponentFrom + 1)`                           
+        //         and `2^exponentTo`                             
+        
+        // Compute the distance between `2^(exponentFrom + 1)`
+        // and `2^exponentTo`.
+        
+        assert(exponentFrom < exponentTo);
+        
+        if (exponentTo - exponentFrom > 1) {
+            // Do every in-between exponents represent normal numbers?
+            if (exponentFrom + 1 >= Traits::smallestNormalNumberExponent) {
+                out += Traits::amountOfNormalNumbersPerExponent
+                       * (exponentTo - exponentFrom - 1);
+            }
+            // Do every in-between exponents represent subnormal numbers?
+            else if (exponentTo <= Traits::smallestNormalNumberExponent) {
+                int exponent = exponentFrom + 1;
+                while (exponent < exponentTo) {
+                    out += std::exp2(
+                        T(exponent - Traits::smallestSubnormalNumberExponent)
+                    );
+                    
+                    ++exponent;
+                }
+            }
[…]
