+++ examples/kokkos/complex/example_division_test.cpp
@@ -1,8 +1,12 @@
+#include <random>
+
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
 #include "Kokkos_Core.hpp"
+#include "Kokkos_Random.hpp"
 
+#include "examples/kokkos/complex/distance.hpp"
 #include "examples/kokkos/complex/example_division.hpp"
 
 namespace reprospect::examples::kokkos::complex {
@@ -12,113 +16,332 @@ using scomplex_t = std::complex<double>;
 
 constexpr double inf = std::numeric_limits<double>::infinity();
 
-// Matcher that handles NaN, signed inf, and approximate equality.
-MATCHER_P(ComplexEq, expected,
-           "complex value near (" + std::to_string(expected.real()) + ", " +
-               std::to_string(expected.imag()) + ")") {
-  auto match_scalar = [&](const double got, const double exp) -> bool {
-    if (std::isnan(exp)) return std::isnan(got);
-    if (std::isinf(exp)) return std::isinf(got) && (got > 0) == (exp > 0);
-    return got == exp;
-  };
-  return match_scalar(arg.real(), expected.real()) &&
-         match_scalar(arg.imag(), expected.imag());
+template <typename T>
+std::string get_msg(const T& value) {
+    std::ostringstream oss;
+    oss << std::scientific << std::setprecision(25);
+    oss << "complex value near (" << value << ")";
+    return oss.str();
 }
 
-//! If @ref expt_std is @c std::nullopt, results for @c std::complex are expected to agree with @ref expt_iec559.
+// Matcher that handles NaN, signed inf, and approximate equality.
+MATCHER_P(ComplexEq, expected, get_msg(expected)) {
+    auto match_scalar = [&](const double got, const double exp) -> bool {
+        if (std::isnan(exp))
+            return std::isnan(got);
+        if (std::isinf(exp))
+            return std::isinf(got) && (got > 0) == (exp > 0);
+        return got == exp;
+    };
+    std::cout << "Comparing " << arg << " with expected " << expected << std::endl;
+    return match_scalar(arg.real(), expected.real()) && match_scalar(arg.imag(), expected.imag());
+}
+
+//! If @ref expt_std is @c std::nullopt, results for @c std::complex are expected to agree with @ref expt_iec559. Similar for @ref expt_scaling.
 struct DivisionCase {
     std::string_view label;
     kcomplex_t num;
     kcomplex_t den;
     kcomplex_t expt_iec559;
     std::optional<kcomplex_t> expt_std = std::nullopt;
+    std::optional<kcomplex_t> expt_scaling = std::nullopt;
 };
 
-class ComplexDivisionTest : public ::testing::TestWithParam<DivisionCase> {};
+class ComplexDivisionTest : public ::testing::TestWithParam<DivisionCase> { };
 
+// like cuda cccl
 TEST_P(ComplexDivisionTest, Iec559) {
-  const auto& p = GetParam();
-  const kcomplex_t result = reprospect::examples::kokkos::complex::iec559(p.num, p.den);
-  EXPECT_THAT(result, ComplexEq(p.expt_iec559)) << p.label;
+    std::cout << std::scientific << std::setprecision(50);
+    const auto& p = GetParam();
+    const kcomplex_t result = reprospect::examples::kokkos::complex::iec559(p.num, p.den);
+    EXPECT_THAT(result, ComplexEq(p.expt_iec559)) << p.label;
+}
+
+// like old kokkos
+TEST_P(ComplexDivisionTest, Scaling) {
[…]
