+++ examples/kokkos/complex/example_division.hpp
@@ -20,12 +20,16 @@ KOKKOS_FUNCTION Kokkos::complex<T> iec559(const Kokkos::complex<T>& x, const Kok
     T __logbw = Kokkos::logb(Kokkos::fmax(Kokkos::fabs(__c), Kokkos::fabs(__d)));
     if (Kokkos::isfinite(__logbw)) {
         __ilogbw = static_cast<int>(__logbw);
+        __a = scalbn(__a, -__ilogbw);
+        __b = scalbn(__b, -__ilogbw);
         __c = scalbn(__c, -__ilogbw);
         __d = scalbn(__d, -__ilogbw);
     }
     T __denom = __c * __c + __d * __d;
-    T __x = scalbn((__a * __c + __b * __d) / __denom, -__ilogbw);
-    T __y = scalbn((__b * __c - __a * __d) / __denom, -__ilogbw);
+    // T __x = scalbn((__a * __c + __b * __d) / __denom, -__ilogbw);
+    // T __y = scalbn((__b * __c - __a * __d) / __denom, -__ilogbw);
+    T __x = (__a * __c + __b * __d) / __denom;
+    T __y = (__b * __c - __a * __d) / __denom;
     if constexpr (EnableBranching) {
         if (Kokkos::isnan(__x) && Kokkos::isnan(__y)) {
             if ((__denom == T(0)) && (!Kokkos::isnan(__a) || !Kokkos::isnan(__b))) {
