+++ .github/workflows/spelling.yml
@@ -16,6 +16,6 @@ jobs:
             image: python:3.12
         steps:
             - uses: actions/checkout@v6
-            - uses: crate-ci/typos@v1.43.5
+            - uses: crate-ci/typos@v1.44.0
               with:
                   config: .github/config/typos.toml
