+++ examples/kokkos/complex/example_division_benchmarking.cpp
@@ -27,6 +27,18 @@ class Division : public ::benchmark::Fixture {
 
     using random_t = Kokkos::Random_XorShift64_Pool<execution_space>;
 
+    template <typename Divisor>
+    struct JustDivide {
+        typename complex_view_t::const_type src_a, src_b;
+        complex_view_t dst_c;
+
+        template <std::integral T>
+        KOKKOS_FUNCTION
+        void operator()(const T index) const noexcept {
+            dst_c(index) = Divisor{}(src_a(index), src_b(index));
+        }
+    };
+
    public:
     FIXME_PARTIAL_OVERRIDE_WARNING_NVCC(TearDown)
     FIXME_PARTIAL_OVERRIDE_WARNING_NVCC(SetUp)
@@ -59,7 +71,7 @@ class Division : public ::benchmark::Fixture {
         for(unsigned short irepeat = 0; irepeat < repeat; ++irepeat) {
         Kokkos::parallel_for(
             Kokkos::RangePolicy(*exec, 0, state.range(range_size)),
-            Scaling<true, complex_view_t>{.src_a = src_a, .src_b = src_b, .dst_c = dst_c});
+            JustDivide<DivisorScalingBranch>{.src_a = src_a, .src_b = src_b, .dst_c = dst_c});
         }
         exec->fence();
     }
@@ -68,7 +80,7 @@ class Division : public ::benchmark::Fixture {
         for(unsigned short irepeat = 0; irepeat < repeat; ++irepeat) {
         Kokkos::parallel_for(
             Kokkos::RangePolicy(*exec, 0, state.range(range_size)),
-            Scaling<false, complex_view_t>{.src_a = src_a, .src_b = src_b, .dst_c = dst_c});
+            JustDivide<DivisorScaling>{.src_a = src_a, .src_b = src_b, .dst_c = dst_c});
         }
         exec->fence();
     }
@@ -76,7 +88,7 @@ class Division : public ::benchmark::Fixture {
     void iec559(const ::benchmark::State& state) const {
         for(unsigned short irepeat = 0; irepeat < repeat; ++irepeat) {
         Kokkos::parallel_for(
-            Kokkos::RangePolicy(*exec, 0, state.range(range_size)), Iec559<complex_view_t>{.src_a = src_a, .src_b = src_b, .dst_c = dst_c});
+            Kokkos::RangePolicy(*exec, 0, state.range(range_size)), JustDivide<DivisorIec559>{.src_a = src_a, .src_b = src_b, .dst_c = dst_c});
         }
         exec->fence();
     }
@@ -84,6 +96,95 @@ class Division : public ::benchmark::Fixture {
    private:
     std::optional<Kokkos::Cuda> exec = std::nullopt;
     complex_view_t src_a, src_b, dst_c;
+};
+
+template <typename Divisor>
+class NewtonFractal : public ::benchmark::Fixture {
+public:
+    using complex_t = Kokkos::complex<double>;
+    using view_t = Kokkos::View<unsigned int**, Kokkos::CudaSpace>;
+    using roots_t = Kokkos::View<complex_t[3], Kokkos::CudaSpace>;
+
+    static constexpr unsigned short range_width = 0; //! Number of samples along the x axis.
+    static constexpr unsigned short range_height = 0; //! Number of samples along the y axis.
+
+    //! Fuction, *e.g.* @c z^3-1.
+    KOKKOS_FUNCTION
+    static constexpr complex_t function(const complex_t z) {
+        return Kokkos::pow(z, 3) - complex_t{1, 0};
+    }
+
+    //! Derivative, *e.g.* @c 3*z^2.
+    KOKKOS_FUNCTION
+    static constexpr complex_t derivative(const complex_t z) {
+        return 3 * z * z;
+    }
+
[…]
