+++ test_load.py
@@ -1,82 +1,92 @@
-import numpy as np
-from matplotlib import pyplot as plt
+import argparse
+import logging
+import pathlib
 
-colors_f = 'colors.bin'
-iterations_f = 'iterations.bin'
+import numpy
+import matplotlib.pyplot
 
-# ---------- Load colors ----------
-with open(colors_f, "rb") as f:
-    dim0 = np.fromfile(f, dtype=np.uint64, count=1)[0]
-    dim1 = np.fromfile(f, dtype=np.uint64, count=1)[0]
-    data = np.fromfile(f, dtype=np.uint32)
+def main(colors: pathlib.Path, iterations: pathlib.Path, output:pathlib.Path) -> None:
+    """
+    Read data from `colors` and `iterations`, assuming they contain 2 sizes of type
+    :py:class:numpy.uint64` and flattened values of type :py:class:`numpy.uint32`.
+    """
+    with colors.open('rb') as fin:
+        dim_0 = numpy.fromfile(fin, dtype=numpy.uint64, count=1)[0]
+        dim_1 = numpy.fromfile(fin, dtype=numpy.uint64, count=1)[0]
+        colors_data = numpy.fromfile(fin, dtype=numpy.uint32).reshape((dim_1, dim_0))
 
-data = data.reshape((dim1, dim0))  # assumes LayoutRight
+    with iterations.open('rb') as fin:
+        dim_0 = numpy.fromfile(fin, dtype=numpy.uint64, count=1)[0]
+        dim_0 = numpy.fromfile(fin, dtype=numpy.uint64, count=1)[0]
+        iterations_data = numpy.fromfile(fin, dtype=numpy.uint32).reshape((dim_1, dim_0))
 
-# ---------- Load iterations ----------
-with open(iterations_f, "rb") as f:
-    dim0_i = np.fromfile(f, dtype=np.uint64, count=1)[0]
-    dim1_i = np.fromfile(f, dtype=np.uint64, count=1)[0]
-    iterations = np.fromfile(f, dtype=np.uint32)
+    assert colors_data.shape == iterations_data.shape
 
-iterations = iterations.reshape((dim1_i, dim0_i))
+    # Convert to float for proper normalization
+    iterations = iterations.astype(np.float64)
 
-print(np.min(iterations))
-print(np.max(iterations))
+    # ---------- Plot ----------
+    fig, axes = plt.subplots(nrows=1, ncols=2,
+                            figsize=(20, 10),
+                            constrained_layout=True)
 
-# Convert to float for proper normalization
-iterations = iterations.astype(np.float64)
+    fig.patch.set_facecolor('black')
 
-# ---------- Plot ----------
-fig, axes = plt.subplots(nrows=1, ncols=2,
-                         figsize=(20, 10),
-                         constrained_layout=True)
+    # First image (categorical colors)
+    im0 = axes[0].imshow(
+        data,
+        origin='lower',
+        extent=[-2, 2, -2, 2]
+    )
 
-fig.patch.set_facecolor('black')
+    axes[0].set_title("Root Index", color="white")
 
-# First image (categorical colors)
-im0 = axes[0].imshow(
-    data,
-    origin='lower',
-    extent=[-2, 2, -2, 2]
-)
-
-axes[0].set_title("Root Index", color="white")
-
-# Add colorbar for first image
-cbar0 = fig.colorbar(im0, ax=axes[0])
-cbar0.ax.yaxis.set_tick_params(color='white')
-plt.setp(cbar0.ax.get_yticklabels(), color='white')
+    # Add colorbar for first image
+    cbar0 = fig.colorbar(im0, ax=axes[0])
+    cbar0.ax.yaxis.set_tick_params(color='white')
[…]
