### 2026-03-04 5e68545
examples(kokkos): generalize benchmarking
We'll have another benchmarking case soon.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
