+++ examples/kokkos/complex/example_division.hpp
@@ -51,12 +51,46 @@ KOKKOS_FUNCTION constexpr Kokkos::complex<RealType>
     return {real, imag};
 }
 
+//! @todo Use @c Kokkos::norm when available.
+template <typename RealType>
+KOKKOS_FUNCTION constexpr RealType norm(const Kokkos::complex<RealType>& value) {
+    return value.real() * value.real() + value.imag() * value.imag();
+}
+
+/**
+ * Adapted from https://github.com/kokkos/kokkos/blob/9174877d49528ab293b6c7f4c6bd932a429b200a/core/src/Kokkos_Complex.hpp#L182-L206.
+ * Similar to https://github.com/NVIDIA/thrust/blob/756c5afc0750f1413da05bd2b6505180e84c53d4/thrust/detail/complex/arithmetic.h#L119.
+ */
+template <bool EnableBranching, typename RealType>
+KOKKOS_FUNCTION constexpr Kokkos::complex<RealType>
+    scaling(const Kokkos::complex<RealType>& x, const Kokkos::complex<RealType>& y) noexcept {
+    const RealType scale = Kokkos::fabs(y.real()) + Kokkos::fabs(y.imag());
+    if constexpr (EnableBranching) {
+        if (scale == RealType(0)) {
+            return {x.real() / scale, x.imag() / scale};
+        }
+    }
+    const Kokkos::complex<RealType> x_scaled = x / scale;
+    const Kokkos::complex<RealType> y_conj_scaled = Kokkos::conj(y) / scale;
+    const RealType y_conj_scaled_norm = norm(y_conj_scaled);
+    return (x_scaled * y_conj_scaled) / y_conj_scaled_norm;
+}
+
 //! Adaptor for @ref logb_scalbn.
 template <bool EnableBranching>
 struct DivisorLogbScalbn {
     template <typename... Args>
     KOKKOS_FUNCTION constexpr auto operator()(Args&&... args) const {
         return logb_scalbn<EnableBranching>(std::forward<Args>(args)...);
+    }
+};
+
+//! Adaptor for @ref scaling.
+template <bool EnableBranching>
+struct DivisorScaling {
+    template <typename... Args>
+    KOKKOS_FUNCTION constexpr auto operator()(Args&&... args) const {
+        return scaling<EnableBranching>(std::forward<Args>(args)...);
     }
 };
 
