+++ docs/source/examples/kokkos/complex/division.rst
@@ -20,4 +20,5 @@ per iteration per pixel.
 
     Newton fractal for :math:`z^4 - 1` with relaxation factor :math:`a = 1`.
 
+.. automodule:: examples.kokkos.complex.example_division_benchmarking
 .. automodule:: examples.kokkos.complex.example_division_plot
