+++ examples/kokkos/complex/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_one_kokkos_example(alignment)
 add_one_kokkos_example(division_plot)
+add_one_kokkos_example_benchmarking(division)
