+++ examples/kokkos/complex/example_division_benchmarking.cpp
@@ -65,7 +65,7 @@ REPROSPECT_EXAMPLES_KOKKOS_COMPLEX_DIVISION_BENCHMARKING(DivisorLogbScalbn, ILog
 REPROSPECT_EXAMPLES_KOKKOS_COMPLEX_DIVISION_BENCHMARKING(DivisorLogbScalbn, ILogbScalbnCompliant, true, true)
 REPROSPECT_EXAMPLES_KOKKOS_COMPLEX_DIVISION_BENCHMARKING(DivisorLogbScalbn, LogbScalbn, false, false)
 REPROSPECT_EXAMPLES_KOKKOS_COMPLEX_DIVISION_BENCHMARKING(DivisorLogbScalbn, LogbScalbnCompliant, true, false)
-REPROSPECT_EXAMPLES_KOKKOS_COMPLEX_DIVISION_BENCHMARKING(DivisorScaling, Scaling, false)
-REPROSPECT_EXAMPLES_KOKKOS_COMPLEX_DIVISION_BENCHMARKING(DivisorScaling, ScalingBranch, true)
+REPROSPECT_EXAMPLES_KOKKOS_COMPLEX_DIVISION_BENCHMARKING(DivisorNormDivision, NormDivision, false)
+REPROSPECT_EXAMPLES_KOKKOS_COMPLEX_DIVISION_BENCHMARKING(DivisorNormDivision, NormDivisionBranch, true)
 
 } // namespace reprospect::examples::kokkos::complex
