+++ reprospect/test/sass/matchers/convert_fp_to_int.py
@@ -43,8 +43,8 @@ class ConvertFpToInt(InstructionMatcher):
         src: str = Register.REG,
     ) -> None:
         """
-        :param src_type: Type of the source floating-point value.
-        :param dst_type: Type of the destination integer value.
+        :param src_dtype: Type of the source floating-point value.
+        :param dst_dtype: Type of the destination integer value.
         """
         src_dtype = TypeInfo.normalize(dtype=src_dtype or 32)
         dst_dtype = TypeInfo.normalize(dtype=dst_dtype or 32)
