+++ examples/kokkos/complex/example_division.py
@@ -1,0 +1,282 @@
+import logging
+import pathlib
+import re
+import sys
+import typing
+
+import numpy
+import pytest
+
+from reprospect.test import environment
+from reprospect.test.case import CMakeAwareTestCase
+from reprospect.test.sass.composite import findall
+from reprospect.test.sass.instruction import OpcodeModsMatcher
+from reprospect.test.sass.instruction.register import Register
+from reprospect.test.sass.matchers.convert_fp_to_int import ConvertFpToInt
+from reprospect.test.sass.matchers.convert_int_to_fp import ConvertIntToFp
+from reprospect.tools.binaries import (
+    CuObjDump,
+    DetailedRegisterUsage,
+    Function,
+    NVDisasm,
+)
+from reprospect.tools.ncu import (
+    Cacher,
+    Command,
+    Metric,
+    MetricCorrelation,
+    MetricCounter,
+    MetricCounterRollUp,
+    ProfilingMetrics,
+    Report,
+)
+from reprospect.tools.sass import Decoder
+from reprospect.tools.sass.decode import RegisterType
+from reprospect.utils import detect
+
+if sys.version_info >= (3, 11):
+    from enum import StrEnum
+else:
+    from backports.strenum.strenum import StrEnum
+
+if sys.version_info >= (3, 12):
+    from typing import override
+else:
+    from typing_extensions import override
+
+class Method(StrEnum):
+    LogbScalbn = 'LogbScalbn'
+    ILogbScalbn = 'ILogbScalbn'
+    Scaling = 'Scaling'
+
+class TestDivision(CMakeAwareTestCase):
+    KOKKOS_TOOLS_NVTX_CONNECTOR_LIB = environment.EnvironmentField(converter=pathlib.Path)
+    """Used in :py:meth:`TestNCU.report`."""
+
+    @classmethod
+    @override
+    def get_target_name(cls) -> str:
+        return 'examples_kokkos_complex_division'
+
+class TestSASS(TestDivision):
+    """
+    Binary analysis.
+    """
+    SIGNATURE: typing.Final[dict[Method, re.Pattern[str]]] = {
+        Method.LogbScalbn:  re.compile(r'reprospect::examples::kokkos::complex::DivisorLogbScalbn<(\(bool\)1|true), (\(bool\)0|false)>'),
+        Method.ILogbScalbn: re.compile(r'reprospect::examples::kokkos::complex::DivisorLogbScalbn<(\(bool\)1|true), (\(bool\)1|true)>'),
+        Method.Scaling:     re.compile(r'reprospect::examples::kokkos::complex::DivisorScaling<(\(bool\)1|true)>'),
+    }
+
+    @property
+    def cubin(self) -> pathlib.Path:
+        return self.cwd / f'{self.get_target_name()}.1.{self.arch.as_sm}.cubin'
+
+    @pytest.fixture(scope='class')
+    def cuobjdump(self) -> CuObjDump:
+        return CuObjDump.extract(file=self.executable, arch=self.arch, sass=True, cwd=self.cwd, cubin=self.cubin.name, demangler=self.demangler)[0]
+
+    @pytest.fixture(scope='class')
+    def nvdisasm(self, cuobjdump: CuObjDump) -> NVDisasm:
+        return NVDisasm(file=cuobjdump.file, arch=self.arch)
+
+    @pytest.fixture(scope='class')
+    def function(self, cuobjdump: CuObjDump) -> dict[Method, Function]:
+        logging.info(list(cuobjdump.functions.keys()))
+        def get_function(method: Method) -> Function:
+            pattern = self.SIGNATURE[method]
[…]
