### 2026-03-05 b9ef95d
examples(kokkos): complex division with ilogb
For even better performance.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
