+++ docs/source/references.bib
@@ -14,6 +14,13 @@
     place = {Houston, TX}
 }
 
+@misc{cppreference-ilogb,
+    title = {std::ilogb, std::ilogbf, std::ilogbl},
+    author = {{cppreference.com contributors}},
+    howpublished = {\url{https://en.cppreference.com/w/cpp/numeric/math/ilogb.html}},
+    note = {Accessed: 2026-03-05}
+}
+
 @article{ctrott-2022,
     author = {{Trott}, Christian R. and {Lebrun-Grandi{\'e}}, Damien and {Arndt}, Daniel and {Ciesko}, Jan and {Dang}, Vinh and {Ellingwood}, Nathan and {Gayatri}, Rahulkumar and {Harvey}, Evan and {Hollman}, Daisy S. and {Ibanez}, Dan and {Liber}, Nevin and {Madsen}, Jonathan and {Miles}, Jeff and {Poliakoff}, David and {Powell}, Amy and {Rajamanickam}, Sivasankaran and {Simberg}, Mikael and {Sunderland}, Dan and {Turcksin}, Bruno and {Wilke}, Jeremiah},
     title = {Kokkos 3: Programming Model Extensions for the Exascale Era},
