### 2026-03-06 38fd879
utils(types): enhance API with a custom `TypeInfo`
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
