### 2026-03-06 12248cc
test(sass): convert from integer to floating-point
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
