+++ docs/source/tests/test/sass/matchers.rst
@@ -4,6 +4,7 @@ Matchers
 .. automodule:: tests.test.sass.matchers.test_add_int128
 .. automodule:: tests.test.sass.matchers.test_add_int32
 .. automodule:: tests.test.sass.matchers.test_cas
+.. automodule:: tests.test.sass.matchers.test_convert_fp_to_int
 .. automodule:: tests.test.sass.matchers.test_convert_fp32_to_fp16
 .. automodule:: tests.test.sass.matchers.test_convert_int_to_fp
 .. automodule:: tests.test.sass.matchers.test_move32
