+++ examples/cuda/virtual_functions/example_dispatch.py
@@ -207,6 +207,9 @@ class TestResourceUsage(TestBinaryAnalysis):
             case 75:
                 expt_static  = {RegisterType.GPR: (6, 5),   RegisterType.PRED: (1, 1), RegisterType.UGPR: (6, 2)}
                 expt_dynamic = {RegisterType.GPR: (24, 18), RegisterType.PRED: (1, 1), RegisterType.UGPR: (6, 2)}
+            case 80:
+                expt_static  = {RegisterType.GPR: (8, 7),   RegisterType.PRED: (1, 1), RegisterType.UGPR: (6, 2)}
+                expt_dynamic = {RegisterType.GPR: (24, 18), RegisterType.PRED: (1, 1), RegisterType.UGPR: (6, 2)}
             case 86:
                 expt_static  = {RegisterType.GPR: (8, 7),   RegisterType.PRED: (1, 1), RegisterType.UGPR: (6, 2)}
                 expt_dynamic = {RegisterType.GPR: (24, 18), RegisterType.PRED: (2, 2), RegisterType.UGPR: (6, 2), RegisterType.UPRED: (1, 1)}
