+++ .github/workflows/build.yml
@@ -117,7 +117,7 @@ jobs:
                   username: ${{ github.actor }}
                   password: ${{ secrets.GITHUB_TOKEN }}
 
-            - uses: docker/setup-buildx-action@v3
+            - uses: docker/setup-buildx-action@v4
 
             - uses: docker/build-push-action@v6.19.2
               with:
@@ -173,7 +173,7 @@ jobs:
 
             - uses: docker/setup-qemu-action@v3
 
-            - uses: docker/setup-buildx-action@v3
+            - uses: docker/setup-buildx-action@v4
 
             - uses: docker/build-push-action@v6.19.2
               with:
@@ -214,7 +214,7 @@ jobs:
                   username: ${{ github.actor }}
                   password: ${{ secrets.GITHUB_TOKEN }}
 
-            - uses: docker/setup-buildx-action@v3
+            - uses: docker/setup-buildx-action@v4
 
             - uses: docker/build-push-action@v6.19.2
               with:
