+++ .github/workflows/build.yml
@@ -171,7 +171,7 @@ jobs:
                   username: ${{ github.actor }}
                   password: ${{ secrets.GITHUB_TOKEN }}
 
-            - uses: docker/setup-qemu-action@v3
+            - uses: docker/setup-qemu-action@v4
 
             - uses: docker/setup-buildx-action@v3
 
