+++ .github/workflows/build.yml
@@ -111,7 +111,7 @@ jobs:
             - uses: actions/checkout@v6
             - run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
 
-            - uses: docker/login-action@v3
+            - uses: docker/login-action@v4
               with:
                   registry: ghcr.io
                   username: ${{ github.actor }}
@@ -165,7 +165,7 @@ jobs:
             - uses: actions/checkout@v6
             - run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
 
-            - uses: docker/login-action@v3
+            - uses: docker/login-action@v4
               with:
                   registry: ghcr.io
                   username: ${{ github.actor }}
@@ -208,7 +208,7 @@ jobs:
             - uses: actions/checkout@v6
             - run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
 
-            - uses: docker/login-action@v3
+            - uses: docker/login-action@v4
               with:
                   registry: ${{ env.REGISTRY }}
                   username: ${{ github.actor }}
