### 2026-03-10 0626967
Update docstring in kokkos complex alignment
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
