### 2026-03-12 5659084
Add random access view trait case in restrict
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
