### 2026-03-12 daef3b0
docker: remove rust
We don't use it yet.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
