### 2026-03-12 06fd6c5
cicd(strategy): cuda 12.8.1 + clang 22
`clang` 21 works fine with `CUDA` 12.8.1.

So let's first use `clang` 22 with `CUDA` 12.8.1. And do a followup later on to bump to `CUDA` 13.

Signed-off-by: Tomasetti Romin <romin.tomasetti@gmail.com>
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
