+++ .github/workflows/strategy.py
@@ -353,6 +353,15 @@ def main(*, args: argparse.Namespace) -> None:
         platforms=('linux/amd64',),
     ), args=args))
 
+    matrix.extend(from_config(Config(
+        cuda_version='12.8.1',
+        ubuntu_version='24.04',
+        python_version='3.12',
+        compilers={'CXX': Compiler(ID='Clang', version='22')},
+        compute_capability=ComputeCapability(major=9, minor=0),
+        platforms=('linux/amd64',),
+    ), args=args))
+
     logging.info(f'Strategy matrix:\n{pprint.pformat(matrix)}')
 
     # All jobs in the matrix build an image.
