+++ .github/workflows/strategy.py
@@ -19,7 +19,7 @@ AVAILABLE_RUNNER_ARCHES: typing.Final[tuple[NVIDIAArch, ...]] = (
 
 SELF_HOSTED: typing.Final[tuple[str, ...]] = ('self-hosted', 'linux', 'docker', 'amd64')
 
-KOKKOS_SHA: typing.Final[str] = '5.0.2'
+KOKKOS_SHA: typing.Final[str] = 'release-candidate-5.1.0'
 
 @dataclasses.dataclass(frozen=False, slots=True)
 class Compiler:
