+++ .github/workflows/build.yml
@@ -231,7 +231,7 @@ jobs:
                       CXX_COMPILER_VERSION=${{ matrix.compilers.CXX.version }}
                       CXX_COMPILER=${{ matrix.compilers.CXX.path }}
                       DOXYGEN_VERSION=1_14_0
-                      GOOGLEBENCHMARK_VERSION=1.9.4
+                      GOOGLEBENCHMARK_VERSION=1.9.5
                       NVTX3_VERSION=3.4.0
                       PYTHON_VERSION=${{ matrix.python_version }}
                   labels: "org.opencontainers.image.source=${{ github.repositoryUrl }}"
