+++ tests/test/sass/test_atomic.py
@@ -11,6 +11,7 @@ import semantic_version
 from reprospect.test.sass.instruction import (
     AtomicMatcher,
     InstructionMatch,
+    ThreadScope,
 )
 from reprospect.test.sass.instruction.memory import MemorySpace
 from reprospect.test.sass.instruction.register import Register
@@ -68,7 +69,7 @@ __global__ void add({type}* __restrict__ const dst, const {type}* __restrict__ c
 __global__ void add({type}* __restrict__ const dst, const {type}* __restrict__ const src)
 {{
     const auto index = blockIdx.x * blockDim.x + threadIdx.x;
-    cuda::atomic_ref<{type}, cuda::thread_scope_device> ref(dst[index]);
+    cuda::atomic_ref<{type}, cuda::thread_scope_{scope}> ref(dst[index]);
     ref.fetch_min(src[index], cuda::memory_order_relaxed);
 }}
 """
@@ -78,6 +79,25 @@ __global__ void exch({type}* __restrict__ const dst, const {type}* __restrict__ 
 {{
     const auto index = blockIdx.x * blockDim.x + threadIdx.x;
     atomicExch(&dst[index], src[index]);
+}}
+"""
+
+    CODE_COMPARE_EXCHANGE = """\
+#include "cuda/atomic"
+
+__global__ void compare_exchange({type}* __restrict__ dst, const {type}* __restrict__ src)
+{{
+    const auto index = blockIdx.x * blockDim.x + threadIdx.x;
+    cuda::atomic_ref<{type}, cuda::thread_scope_{scope}> ref(dst[index]);
+
+    {type} old_val = ref.load(cuda::memory_order_relaxed);
+    {type} new_val;
+
+    do {{
+        new_val = min(old_val, src[index]);
+    }} while (!ref.compare_exchange_{consistency}(old_val, new_val,
+                                       cuda::memory_order_acquire,
+                                       cuda::memory_order_acquire));
 }}
 """
 
@@ -166,10 +186,15 @@ __global__ void atomic_exch_kernel() {
         decoder, _ = get_decoder(cwd=workdir, arch=parameters.arch, file=FILE, cmake_file_api=cmake_file_api)
 
         # Find the atomic CAS.
+        if cmake_file_api.toolchains['CUDA']['compiler']['id'] == 'Clang' and \
+            semantic_version.Version(cmake_file_api.toolchains['CUDA']['compiler']['version']) in semantic_version.SimpleSpec('>21'):
+            expt_scope = 'SYSTEM'
+        else:
+            expt_scope = 'DEVICE'
         _, _, matched = self.match_one(
             decoder=decoder,
             arch=parameters.arch,
-            operation='CAS', consistency='STRONG', scope='DEVICE',
+            operation='CAS', consistency='STRONG', scope=expt_scope,
             dtype=word[0],
         )
 
@@ -329,10 +354,10 @@ __global__ void atomic_exch_kernel() {
 
     def test_min_relaxed_device_int(self, request, workdir, parameters: Parameters, cmake_file_api: cmake.FileAPI):
         """
-        Test with :py:attr:`CODE_MIN` for `int`.
+        Test with :py:attr:`CODE_MIN` for `int` and :py:data:`reprospect.test.sass.instruction.ThreadScope.DEVICE` scope.
         """
         FILE = workdir / f'{request.node.originalname}.{parameters.arch.as_sm}.cu'
-        FILE.write_text(self.CODE_MIN.format(type='int'))
+        FILE.write_text(self.CODE_MIN.format(type='int', scope='device'))
 
         decoder, output = get_decoder(cwd=workdir, arch=parameters.arch, file=FILE, cmake_file_api=cmake_file_api, ptx=True)
 
[…]
