### 2026-03-12 c054d7e
google(benchmark): get rid of internal
Followup of #534.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
