### 2026-03-13 4c90930
cicd(artifacts): add cuda version
To distinguish several same compiler same architecture.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
