### 2026-03-17 55a6e02
tests(cmake-file-api): compiler for toolchain
Instead of relying on a dictionary, let's rely on what cmake-file-api already provides.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
