+++ requirements/requirements.python.txt
@@ -4,7 +4,7 @@ mypy==1.19.1
 myst-nb>=1.3.0
 nvtx>=0.2.13
 pylint>=4.0.1
-ruff==0.15.6
+ruff==0.15.7
 setuptools==82.0.1
 sphinx<=8.2.3
 sphinx-copybutton==0.5.2
