### 2026-03-25 927a86d
mypyc(setup.py): adding a few missing files
Signed-off-by: Tomasetti Romin <romin.tomasetti@gmail.com>
