+++ setup.py
@@ -28,6 +28,7 @@ def enable_mypyc(dist: Distribution) -> None:
             'reprospect/test/sass/instruction/constant.py',
             'reprospect/test/sass/instruction/floating.py',
             'reprospect/test/sass/instruction/half.py',
+            'reprospect/test/sass/instruction/immediate.py',
             'reprospect/test/sass/instruction/instruction.py',
             'reprospect/test/sass/instruction/integer.py',
             'reprospect/test/sass/instruction/load.py',
@@ -57,9 +58,11 @@ def enable_mypyc(dist: Distribution) -> None:
             'reprospect/tools/sass/controlflow.py',
             'reprospect/tools/sass/decode.py',
             'reprospect/utils/cmake.py',
+            'reprospect/utils/compile_command.py',
             'reprospect/utils/detect.py',
             'reprospect/utils/ldd.py',
             'reprospect/utils/nvcc.py',
+            'reprospect/utils/rich_helpers.py',
             'reprospect/utils/subprocess_helpers.py',
             'reprospect/utils/types.py',
         ],
