### 2026-03-25 09fd506
style(ruff): more checks
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
