+++ pyproject.toml
@@ -149,7 +149,10 @@ extend-select = [
     "I",
     "INT",
     "ISC",
+    "NPY001",
     "NPY002",
+    "NPY003",
+    "PD",
     "PERF",
     "PGH",
     "PIE",
@@ -179,15 +182,34 @@ extend-select = [
     "TRY002",
     "TRY004",
     "TRY2",
+    "TRY3",
     "TRY4",
+    "UP001",
+    "UP003",
+    "UP004",
+    "UP005",
     "UP006",
     "UP007",
+    "UP008",
     "UP009",
+    "UP010",
+    "UP011",
     "UP012",
+    "UP013",
+    "UP014",
     "UP015",
     "UP017",
+    "UP018",
+    "UP019",
     "UP020",
+    "UP021",
+    "UP022",
+    "UP023",
+    "UP024",
+    "UP025",
+    "UP026",
     "UP028",
+    "UP029",
     "UP030",
     "UP031",
     "UP032",
