### 2026-03-26 #10067 @cedricchevalier19
@nkrusch I should have more bandwidth next week to contribute to this review. No hard blockers.
