### 2026-04-01 81b1b6e
Remove 'ruff' from build.yml tools list
Removed 'ruff' from the list of tools in the build workflow.

Signed-off-by: Maarten Arnst <maarten.arnst@ulg.ac.be>
