+++ requirements/requirements.python.txt
@@ -1,6 +1,6 @@
 coverage==7.13.5
 matplotlib
-mypy==1.19.1
+mypy==1.20.0
 myst-nb>=1.3.0
 nvtx>=0.2.13
 pylint>=4.0.1
