### 2026-04-01 8ebff5f
utils(cmake): code model has a name key/value pair (str -> str)
It also confused mypy.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
