+++ tests/tools/binaries/test_nvdisasm.py
@@ -106,7 +106,7 @@ class TestNVDisasm:
                         RegisterType.PRED: (1, 1),
                         RegisterType.UGPR: (6, 2),
                     }
-                case 90 | 100 | 103 | 120:
+                case 90 | 100 | 103 | 120 | 121:
                     expt_register_usage_details = {
                         RegisterType.GPR:  (8, 7),
                         RegisterType.PRED: (1, 1),
@@ -184,7 +184,7 @@ class TestNVDisasm:
                         self.SYMBOLS[0]: {RegisterType.GPR: (22, 18), RegisterType.PRED: (1, 1), RegisterType.UGPR: (8, 4)},
                         self.SYMBOLS[1]: {RegisterType.GPR: (12, 10), RegisterType.PRED: (1, 1), RegisterType.UGPR: (6, 2)},
                     }
-                case 120:
+                case 120 | 121:
                     expt_register_usage_details = {
                         self.SYMBOLS[0]: {RegisterType.GPR: (22, 18),                            RegisterType.UGPR: (8, 4)},
                         self.SYMBOLS[1]: {RegisterType.GPR: (12, 10), RegisterType.PRED: (1, 1), RegisterType.UGPR: (6, 2)},
