### 2026-04-02 6a608b4
cicd(images): cc121
This will allow to run examples as well.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
