### 2026-04-10 c6830d8
build(deps): bump https://github.com/astral-sh/ruff-pre-commit
Bumps [https://github.com/astral-sh/ruff-pre-commit](https://github.com/astral-sh/ruff-pre-commit) from v0.15.9 to 0.15.10.
- [Release notes](https://github.com/astral-sh/ruff-pre-commit/releases)
- [Commits](https://github.com/astral-sh/ruff-pre-commit/compare/v0.15.9...v0.15.10)

---
updated-dependencies:
- dependency-name: https://github.com/astral-sh/ruff-pre-commit
  dependency-version: 0.15.10
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
