+++ .pre-commit-config.yaml
@@ -9,7 +9,7 @@ repos:
       exclude: '.*\.sass$'
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.15.9
+    rev: v0.15.10
     hooks:
     - id: ruff-check
       args: [--fix, --config=pyproject.toml]
