### 2026-04-15 uliegecsm/reprospect#565 @cedricchevalier19
Improvements for JOSS
Related to above, I think the goal of the software can be clearer. Currently, the documentation overview explains more how it works than what is problem solved.

For JOSS review https://github.com/openjournals/joss-reviews/issues/10067
