### 2026-04-19 c6c178d
docs(sphinx): better left navigation menu
Fixes #572.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
