### 2026-04-19 2668794
docs(sphinx): better left navigation menu
Fixes #572.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
