+++ docs/source/user_guide.rst
@@ -1,0 +1,13 @@
+User Guide
+==========
+
+This guide walks you through installing and using `ReProspect`,
+covering basic usage, key analysis features, and caching of API tracing and kernel profiling results.
+
+.. toctree::
+   :maxdepth: 4
+
+   install
+   getting-started
+   analysis
+   caching
