### 2026-04-21 a13ca37
tests(assets): remove confusing `test_` prefix
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
