+++ tests/test/sass/controlflow/test_block.py
@@ -21,7 +21,7 @@ from tests.test.sass.test_instruction import get_compilation_output
 
 
 class TestBasicBlock:
-    FILE: typing.Final[pathlib.Path] = pathlib.Path(__file__).parent.parent.parent.parent / 'assets' / 'test_half.cu'
+    FILE: typing.Final[pathlib.Path] = pathlib.Path(__file__).parent.parent.parent.parent / 'assets' / 'half.cu'
 
     SIGNATURE_INDIVIDUAL: typing.Final[str] = 'pow2_individual(__half *, const __half *, unsigned int)'
     SIGNATURE_PACKED: typing.Final[str] = 'pow2_packed(__half *, const __half *, unsigned int)'
