### 2026-04-21 b350dd9
tests: add "integration" directory for "more involved" tests
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
