+++ docs/source/tests/integration.rst
@@ -1,0 +1,8 @@
+Integration tests
+=================
+
+Tests that exercise many features,
+but do not seek to exemplify use cases.
+For use cases, see :ref:`examples`.
+
+.. automodule:: tests.integration.test_half
