### 2026-04-21 37d26fb
tests(saxpy): remove redundant tests
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
