+++ tests/assets/saxpy_nil.cpp
@@ -1,0 +1,7 @@
+#include "saxpy.cu"
+
+int main() {
+    saxpy_kernel<<<1, 1, 0, 0>>>(0, 0, nullptr, nullptr);
+
+    return 0;
+}
