+++ docs/source/tests/test.rst
@@ -19,4 +19,3 @@ Others
 
 .. automodule:: tests.test.test_graph
 .. automodule:: tests.test.test_half
-.. automodule:: tests.test.test_saxpy

[…]
