### 2026-04-21 be10edc
Enable tests on arm64
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
