+++ .github/workflows/strategy.py
@@ -1,6 +1,9 @@
+from __future__ import annotations
+
 import argparse
 import copy
 import dataclasses
+import enum
 import json
 import logging
 import pprint
@@ -10,14 +13,44 @@ from reprospect.tools.architecture import ComputeCapability, NVIDIAArch
 
 COMPILER_ID = typing.Literal['Clang', 'GNU', 'NVIDIA']
 LANGUAGE = typing.Literal['CUDA', 'CXX']
-PLATFORM = typing.Literal['linux/amd64', 'linux/arm64']
 
-AVAILABLE_RUNNER_ARCHES: typing.Final[tuple[NVIDIAArch, ...]] = (
-    NVIDIAArch.from_str('VOLTA70'),
-    NVIDIAArch.from_str('BLACKWELL120'),
+class OS(enum.StrEnum):
+    LINUX = 'linux'
+
+class Architecture(enum.StrEnum):
+    AMD64 = 'amd64'
+    ARM64 = 'arm64'
+
+@dataclasses.dataclass(frozen=True, slots=True)
+class Platform:
+    """
+    Inspired by https://docs.docker.com/reference/cli/docker/manifest/#inspect-an-images-manifest-and-get-the-osarch-info.
+    """
+    os: OS
+    architecture: Architecture
+
+    def __repr__(self) -> str:
+        return f'{self.os}/{self.architecture}'
+
+    @classmethod
+    def from_str(cls, value: str) -> Platform:
+        os, architecture = value.split('/')
+        return Platform(os=OS(os), architecture=Architecture(architecture))
+
+@dataclasses.dataclass(frozen=True, slots=True)
+class Runner:
+    """
+    Describe a runner configuration.
+    """
+    platform: Platform
+    arch: NVIDIAArch
+
+AVAILABLE_RUNNERS: typing.Final[tuple[Runner, ...]] = (
+    Runner(platform=Platform(os=OS('linux'), architecture=Architecture('amd64')), arch=NVIDIAArch.from_str('VOLTA70')),
+    Runner(platform=Platform(os=OS('linux'), architecture=Architecture('amd64')), arch=NVIDIAArch.from_str('BLACKWELL120')),
 )
 
-SELF_HOSTED: typing.Final[tuple[str, ...]] = ('self-hosted', 'linux', 'docker', 'amd64')
+SELF_HOSTED: typing.Final[tuple[str, ...]] = ('self-hosted', 'linux', 'docker')
 
 KOKKOS_SHA: typing.Final[str] = '5.1.0'
 
@@ -48,7 +81,7 @@ class Config:
     python_version: str
     compilers: dict[LANGUAGE, Compiler]
     compute_capability: ComputeCapability
-    platforms: tuple[PLATFORM, ...]
+    platforms: tuple[Platform, ...]
     enable_tests: bool | None = None
     enable_examples: bool | None = None
 
@@ -69,13 +102,13 @@ class Config:
         Each platform is a separate job, because multi-arch builds are too slow due to emulation.
         """
         for platform in self.platforms:
-            # We don't run tests and examples for 'linux/arm64'.
-            if platform == 'linux/arm64':
-                if self.enable_tests is True or self.enable_examples is True:
-                    raise RuntimeError(f'Running tests or examples for {platform} is not supported yet.')
-                enable_tests = enable_examples = False
+            enable_tests = self.enable_tests is not False
+            # We don't run examples for 'linux/arm64'.
+            if platform.os == 'linux' and platform.architecture == 'arm64':
+                if self.enable_examples is True:
+                    raise RuntimeError(f'Running examples for {platform} is not supported yet.')
+                enable_examples = False
             else:
-                enable_tests    = self.enable_tests    is not False
[…]
