### 2026-04-22 c26c34c
Enable examples on arm64. Generalize ilogb scaling to arm64.
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
