### 2026-04-23 a7cb13a
docs(tests): better main test page experience
This PR improves the user experience when landing on the main test page.

It has been split in 3 parts:
* tests
* utilities
* assets

to better guide the reader.

Some `.rst` were improved by telling which module they contain tests for.

Closes #576.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
