+++ docs/source/tests/tools.rst
@@ -1,6 +1,8 @@
 Tools
 =====
 
+Tests for :py:mod:`reprospect.tools`.
+
 Architecture
 ------------
 
