### 2026-04-23 1168b91
cicd: cc100 with 13.2 + arm64
Signed-off-by: Tomasetti Romin <romin.tomasetti@gmail.com>
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
