### 2026-04-23 22e1e72
paper(joss): clearly defining what we mean by prospecting
This PR addresses #585 by making more explicit what we mean by prospecting.

The first paragraph of the paper is copy/pasted into the readme, and the readme is included in the Sphinx documentation to fully address #585.

Closes #585.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
