+++ README.md
@@ -12,8 +12,14 @@
 
 [![JOSS](https://joss.theoj.org/papers/31d0b8cb561f2ae10c399e2373543e7f/status.svg)](https://joss.theoj.org/papers/31d0b8cb561f2ae10c399e2373543e7f)
 
-# reprospect
+# ReProspect
 
-`ReProspect` provides a framework for reproducible prospecting of CUDA applications.
+`ReProspect` [/riːˈprɒs.pɛkt/] is a Python framework for prospecting CUDA code,
+designed to ensure reproducibility through a fully programmatic approach.
+Prospecting encompasses three complementary ways of characterizing
+CUDA-based libraries and software components:
+how they interact with the CUDA runtime through API tracing,
+how kernels perform through kernel profiling,
+and how source constructs translate into machine code through binary analysis.
 
 See [the documentation](https://uliegecsm.github.io/reprospect/) for more information.
