### 2026-04-24 e281eb0
Additional references binary analysis
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
