+++ papers/joss/2025/paper.bib
@@ -97,12 +97,14 @@
     pages="79-91"
 }
 
-@misc{mayr-2025,
+@article{mayr2026,
     title = {Trilinos: Enabling Scientific Computing Across Diverse Hardware Architectures at Scale},
     author = {Matthias Mayr and Alexander Heinlein and Christian Glusa and Siva Rajamanickam and Maarten Arnst and Roscoe Bartlett and Luc Berger-Vergiat and Erik Boman and Karen Devine and Graham Harper and Michael Heroux and Mark Hoemmen and Jonathan Hu and Brian Kelley and Kyungjoo Kim and Drew P. Kouri and Paul Kuberry and Kim Liegeois and Curtis C. Ober and Roger Pawlowski and Carl Pearson and Mauro Perego and Eric Phipps and Denis Ridzal and Nathan V. Roberts and Christopher Siefert and Heidi Thornquist and Romin Tomasetti and Christian R. Trott and Raymond S. Tuminaro and James M. Willenbring and Michael M. Wolf and Ichitaro Yamazaki},
-    year = {2025},
-    archivePrefix = {arXiv},
-    doi = {10.48550/arXiv.2503.08126},
+    journal = {ACM Transactions on Mathematical Software},
+    year = {2026},
+    publisher = {Association for Computing Machinery},
+    note = {Just Accepted},
+    doi = {10.1145/3802822},
 }
 
 @misc{ncu2025,
@@ -177,4 +179,40 @@
     doi = {10.1016/j.parco.2021.102837},
     volume = {108},
     pages = {102837},
+}
+
+@inproceedings{he2025,
+    author = {He, Guoliang and Yoneki, Eiko},
+    title = {CuAsmRL: Optimizing GPU SASS Schedules via Deep Reinforcement Learning},
+    booktitle = {Proceedings of the 23rd ACM/IEEE International Symposium on Code Generation and Optimization},
+    year = {2025},
+    publisher = {Association for Computing Machinery},
+    doi = {10.1145/3696443.3708943},
+    pages = {493-506},
+}
+
+@misc{cuassembler2023,
+    author = {Cloudcores},
+    title = {CuAssembler: An unofficial cuda assembler, for all generations of SASS},
+    year = {2023},
+    howpublished = {\url{https://github.com/cloudcores/CuAssembler}},
+    note = {Accessed: 2026-04-24}
+}
+
+@inproceedings{villa2019,
+    author = {Villa, Oreste and Stephenson, Mark and Nellans, David and Keckler, Stephen W.},
+    title = {NVBit: A Dynamic Binary Instrumentation Framework for NVIDIA GPUs},
+    booktitle = {Proceedings of the 52nd Annual IEEE/ACM International Symposium on Microarchitecture},
+    year = {2019},
+    publisher = {Association for Computing Machinery},
+    doi = {10.1145/3352460.3358307},
+    pages = {372-383},
+}
+
+@inproceedings{desclaux2012,
+    author = {Desclaux, Fabrice},
+    title = {Miasm Framework de reverse engineering},
+    booktitle = {Actes de la Conf\'{e}rence francophone sur le th\`{e}me de la s\'{e}curit\'{e} de l'information},
+    year = {2012},
+    howpublished = {\url{https://www.sstic.org/media/SSTIC2012/SSTIC-actes/miasm_framework_de_reverse_engineering/SSTIC2012-Article-miasm_framework_de_reverse_engineering-desclaux_1.pdf}},
 }
