### 2026-04-24 e48509e
docs: add CICD strategy matrix
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
