### 2026-04-24 1f3c0d6
docs(sphinx): improve overview and getting started example
We'd like to better put the SASS matching in the spotlight.

Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
