+++ docs/source/analysis.rst
@@ -6,6 +6,8 @@ Analysis workflow
 Three components
 ----------------
 
+.. _api_tracing:
+
 API tracing
 ~~~~~~~~~~~
 
@@ -13,12 +15,16 @@ API tracing
 * :py:class:`reprospect.tools.nsys.Session`
 * :py:class:`reprospect.tools.nsys.Report`
 
+.. _kernel_profiling:
+
 Kernel profiling
 ~~~~~~~~~~~~~~~~
 
 * :py:class:`reprospect.tools.ncu.Cacher`
 * :py:class:`reprospect.tools.ncu.Session`
 * :py:class:`reprospect.tools.ncu.Report`
+
+.. _binary_analysis:
 
 Binary analysis
 ~~~~~~~~~~~~~~~
