### 2026-04-24 630f027
tools(elf): some more useful `NvInfoEIATTR` with tests
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
