### 2026-04-24 1ff34cb
build(deps): update myst-nb requirement from >=1.3.0 to >=1.4.0
Updates the requirements on [myst-nb](https://github.com/executablebooks/myst-nb) to permit the latest version.
- [Release notes](https://github.com/executablebooks/myst-nb/releases)
- [Changelog](https://github.com/executablebooks/MyST-NB/blob/main/CHANGELOG.md)
- [Commits](https://github.com/executablebooks/myst-nb/compare/v1.3.0...v1.4.0)

---
updated-dependencies:
- dependency-name: myst-nb
  dependency-version: 1.4.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
