+++ requirements/requirements.python.txt
@@ -1,7 +1,7 @@
 coverage==7.13.5
 matplotlib
 mypy==1.20.1
-myst-nb>=1.3.0
+myst-nb>=1.4.0
 nvtx>=0.2.13
 pylint>=4.0.1
 ruff==0.15.11
