+++ requirements/requirements.python.txt
@@ -8,7 +8,7 @@ ruff==0.15.11
 setuptools==82.0.1
 sphinx<=8.2.3
 sphinx-copybutton==0.5.2
-sphinx-rtd-theme>=3.0.2
+sphinx-rtd-theme>=3.1.0
 sphinx-github-style>=1.2.2
 sphinxcontrib-bibtex>=2.6.5
 sphinxcontrib-mermaid==2.0.1
