+++ requirements/requirements.python.txt
@@ -1,6 +1,6 @@
 coverage==7.13.5
 matplotlib
-mypy==1.20.1
+mypy==1.20.2
 myst-nb>=1.4.0
 nvtx>=0.2.13
 pylint>=4.0.1
