### 2026-04-26 93ac424
build(deps): update nvtx requirement from >=0.2.13 to >=0.2.15
Updates the requirements on [nvtx](https://github.com/NVIDIA/NVTX) to permit the latest version.
- [Release notes](https://github.com/NVIDIA/NVTX/releases)
- [Commits](https://github.com/NVIDIA/NVTX/commits)

---
updated-dependencies:
- dependency-name: nvtx
  dependency-version: 0.2.15
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
