+++ requirements/requirements.python.txt
@@ -2,7 +2,7 @@ coverage==7.13.5
 matplotlib
 mypy==1.20.2
 myst-nb>=1.4.0
-nvtx>=0.2.13
+nvtx>=0.2.15
 pylint>=4.0.5
 ruff==0.15.11
 setuptools==82.0.1
