### 2026-04-27 6598ae2
examples(kokkos): static batch size
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
