+++ docs/source/examples/kokkos.rst
@@ -8,4 +8,5 @@
    kokkos/complex
    kokkos/graph
    kokkos/half
+   kokkos/static_batch_size
    kokkos/view
