+++ examples/kokkos/CMakeLists.txt
@@ -27,4 +27,5 @@ add_subdirectory(atomic)
 add_subdirectory(complex)
 add_subdirectory(graph)
 add_subdirectory(half)
+add_subdirectory(static_batch_size)
 add_subdirectory(view)
