### 2026-04-27 22a89fb
cicd(pre-commit): clang-format
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
