+++ .pre-commit-config.yaml
@@ -8,6 +8,12 @@ repos:
     - id: trailing-whitespace
       exclude: '.*\.sass$'
 
+  - repo: https://github.com/pre-commit/mirrors-clang-format
+    rev: v22.1.2
+    hooks:
+    - id: clang-format
+      types_or: [c++, c, cuda]
+
   - repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.15.12
     hooks:
