### 2026-04-27 uliegecsm/reprospect#565 @romintomasetti
Hi @cedricchevalier19 !

Many thanks for the suggestions. I think we addressed all the requested changes.

I'll let close this issue.
