### 2026-04-27 3378ccf
typos: move to pre-commit
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
