+++ examples/kokkos/graph/example_then.cpp
@@ -9,6 +9,7 @@
  */
 
 namespace reprospect::examples::kokkos::graph {
+
 template <typename ViewType>
 struct Functor {
     ViewType data;
