+++ examples/kokkos/graph/CMakeLists.txt
@@ -1,2 +1,3 @@
+add_one_kokkos_example(definition)
 add_one_kokkos_example(dispatch)
 add_one_kokkos_example(then)
