+++ requirements/requirements.python.txt
@@ -4,7 +4,7 @@ mypy==1.20.2
 myst-nb>=1.4.0
 nvtx>=0.2.15
 pylint>=4.0.5
-ruff==0.15.11
+ruff==0.15.12
 setuptools==82.0.1
 sphinx<=8.2.3
 sphinx-copybutton==0.5.2
