### 2026-05-01 8425353
build(deps): bump https://github.com/pre-commit/mirrors-clang-format
Bumps [https://github.com/pre-commit/mirrors-clang-format](https://github.com/pre-commit/mirrors-clang-format) from v22.1.2 to 22.1.4.
- [Commits](https://github.com/pre-commit/mirrors-clang-format/compare/v22.1.2...v22.1.4)

---
updated-dependencies:
- dependency-name: https://github.com/pre-commit/mirrors-clang-format
  dependency-version: 22.1.4
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
