+++ .pre-commit-config.yaml
@@ -9,7 +9,7 @@ repos:
       exclude: '.*\.sass$'
 
   - repo: https://github.com/pre-commit/mirrors-clang-format
-    rev: v22.1.2
+    rev: v22.1.4
     hooks:
     - id: clang-format
       types_or: [c++, c, cuda]
