+++ tests/cublas.py
@@ -40,11 +40,12 @@ class CuBLAS:
             for file in CuObjDump.extract_elf(file=self.libcublas, arch=arch, name=name, cwd=cwd)
         )
 
-    def cubin(self, *, arch: NVIDIAArch, cwd: pathlib.Path, randomly: bool = True, **kwargs) -> CuObjDump:
+    def cuobjdump(self, *, arch: NVIDIAArch, cwd: pathlib.Path, **kwargs) -> CuObjDump:
         """
-        Pick a cubin file from :py:attr:`libcublas` and read it with :py:class:`reprospect.tools.binaries.CuObjDump`.
+        Pick a random cubin file from :py:attr:`libcublas` with :py:meth:`extract`
+        and read it with :py:class:`reprospect.tools.binaries.CuObjDump`.
         """
-        [cubin] = self.extract(arch=arch, cwd=cwd, randomly=randomly)
+        [cubin] = self.extract(arch=arch, cwd=cwd, randomly=True)
         assert cubin.is_file()
 
         return CuObjDump(file=cubin, arch=arch, **kwargs)
