### 2026-05-05 7825d08
tests: match all operands of all instructions of `cuBLAS`
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
