+++ tests/test/sass/instruction/test_cublas.py
@@ -1,0 +1,107 @@
+import dataclasses
+import logging
+import pathlib
+import re
+import typing
+
+import pytest
+import regex
+
+from reprospect.test.sass.instruction.address import AddressMatcher, MemorySpace
+from reprospect.test.sass.instruction.constant import ConstantMatcher
+from reprospect.test.sass.instruction.immediate import Immediate
+from reprospect.test.sass.instruction.instruction import AnyMatcher
+from reprospect.test.sass.instruction.register import RegisterMatcher
+from reprospect.tools.sass.decode import Decoder, Instruction
+from reprospect.utils import cmake
+
+from tests.cublas import CuBLAS
+from tests.parameters import PARAMETERS, Parameters
+
+FIXIT_FILTER: typing.Final[re.Pattern[str]] = re.compile(r'\] \[| ,')
+FIXIT_REPLACEMENTS: typing.Final[dict[str, str]] = {
+    '] [': '][',
+    ' ,': ',',
+}
+
+def fixit(instruction: Instruction) -> Instruction:
+    """
+    Fix the few instructions that do not conform to the expected format.
+
+        +------------+--------+
+        | Before     | After  |
+        +============+========+
+        | ``]␣[``    | ``][`` |
+        +------------+--------+
+        | :code:`␣,` | ``,``  |
+        +------------+--------+
+
+    Here are a few examples:
+
+    * :code:`HADD2.F32 R10, -RZ, c[0x0]␣[0x160].H0_H0` becomes :code:`HADD2.F32 R10, -RZ, c[0x0][0x160].H0_H0`.
+    * :code:`FSETP.GEU.AND P1, PT, |R151|, +INF␣, PT` becomes :code:`FSETP.GEU.AND P1, PT, |R151|, +INF, PT`.
+
+    These cases are rare (a handful out of millions of instructions), so the
+    search acts as a fast-path guard before the substitution.
+    """
+    if FIXIT_FILTER.search(instruction.instruction) is not None:
+        output = FIXIT_FILTER.sub(
+            repl=lambda m: FIXIT_REPLACEMENTS[m.group()],
+            string=instruction.instruction,
+        )
+        return dataclasses.replace(instruction, instruction=output)
+    return instruction
+
+class TestCuBLAS:
+    @pytest.mark.parametrize('parameters', PARAMETERS, ids=str)
+    def test_operands(self, workdir: pathlib.Path, parameters: Parameters, cmake_file_api: cmake.FileAPI) -> None:
+        """
+        For each supported architecture, pick one cubin from cuBLAS.
+        Then, loop over all instructions of each function, and match each operand.
+        """
+        cublas = CuBLAS(cmake_file_api=cmake_file_api)
+        try:
+            cuobjdump = cublas.cuobjdump(arch=parameters.arch, cwd=workdir, sass=True)
+        except IndexError:
+            pytest.skip(f'The library {cublas.libcublas} does not contain any CUDA binary file for {parameters.arch}.')
+
+        if len(cuobjdump.functions) == 0:
+            pytest.skip(f'{cuobjdump.file} does not contain any function for {parameters.arch}.')
+
+        # Match any instruction.
+        matcher = AnyMatcher()
+
+        matchers_op: list = []
+
+        # Matchers for operands that are well covered.
+        matchers_op.extend((
+            regex.compile(AddressMatcher.build_pattern(arch=parameters.arch, memory=MemorySpace.GENERIC, captured=False)),
+            regex.compile(AddressMatcher.build_pattern(arch=parameters.arch, memory=MemorySpace.GLOBAL, captured=False)),
+            regex.compile(AddressMatcher.build_pattern(arch=parameters.arch, memory=MemorySpace.LOCAL, captured=False)),
[…]
