+++ docs/source/tests/test/sass/instruction.rst
@@ -3,6 +3,7 @@ Instruction
 
 .. automodule:: tests.test.sass.instruction.test_address
 .. automodule:: tests.test.sass.instruction.test_constant
+.. automodule:: tests.test.sass.instruction.test_cublas
 .. automodule:: tests.test.sass.instruction.test_fp32add
 .. automodule:: tests.test.sass.instruction.test_fp64add
 .. automodule:: tests.test.sass.instruction.test_half
