### 2026-05-08 009098c
build(deps): update sphinxcontrib-bibtex requirement
Updates the requirements on [sphinxcontrib-bibtex](https://github.com/mcmtroffaes/sphinxcontrib-bibtex) to permit the latest version.
- [Changelog](https://github.com/mcmtroffaes/sphinxcontrib-bibtex/blob/develop/CHANGELOG.rst)
- [Commits](https://github.com/mcmtroffaes/sphinxcontrib-bibtex/compare/2.6.5...2.7.0)

---
updated-dependencies:
- dependency-name: sphinxcontrib-bibtex
  dependency-version: 2.7.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
