+++ requirements/requirements.python.txt
@@ -10,7 +10,7 @@ sphinx<=8.2.3
 sphinx-copybutton==0.5.2
 sphinx-rtd-theme>=3.1.0
 sphinx-github-style>=1.2.2
-sphinxcontrib-bibtex>=2.6.5
+sphinxcontrib-bibtex>=2.7.0
 sphinxcontrib-mermaid==2.0.1
 sphinxcontrib-tikz>=0.4.20
 typeguard
