### 2026-05-08 cbfb30f
build(deps): bump https://github.com/crate-ci/typos
Bumps [https://github.com/crate-ci/typos](https://github.com/crate-ci/typos) from v1.46.0 to 1.46.1.
- [Release notes](https://github.com/crate-ci/typos/releases)
- [Changelog](https://github.com/crate-ci/typos/blob/master/CHANGELOG.md)
- [Commits](https://github.com/crate-ci/typos/compare/v1.46.0...v1.46.1)

---
updated-dependencies:
- dependency-name: https://github.com/crate-ci/typos
  dependency-version: 1.46.1
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
