+++ requirements/requirements.python.txt
@@ -11,7 +11,7 @@ sphinx-copybutton==0.5.2
 sphinx-rtd-theme>=3.1.0
 sphinx-github-style>=1.2.2
 sphinxcontrib-bibtex>=2.7.0
-sphinxcontrib-mermaid==2.0.1
+sphinxcontrib-mermaid==2.0.2
 sphinxcontrib-tikz>=0.4.20
 typeguard
 types-regex
