+++ requirements/requirements.python.txt
@@ -1,6 +1,6 @@
 coverage==7.13.5
 matplotlib
-mypy==1.20.2
+mypy==2.0.0
 myst-nb>=1.4.0
 nvtx>=0.2.15
 pylint>=4.0.5
