+++ pyproject.toml
@@ -1,7 +1,7 @@
 [build-system]
 requires = [
     "setuptools",
-    "mypy[mypyc]==1.20.2",
+    "mypy[mypyc]==2.0.0",
 
     # mypyc requires that runtime dependencies are also installed
     # during the build, since it parses the code base.
