+++ .github/workflows/build.yml
@@ -125,7 +125,7 @@ jobs:
 
             - uses: docker/setup-buildx-action@v4
 
-            - uses: docker/build-push-action@v7.1.0
+            - uses: docker/build-push-action@v7.2.0
               with:
                   context: .
                   platforms: linux/amd64
@@ -180,7 +180,7 @@ jobs:
 
             - uses: docker/setup-buildx-action@v4
 
-            - uses: docker/build-push-action@v7.1.0
+            - uses: docker/build-push-action@v7.2.0
               with:
                   context: null
                   platforms: linux/amd64,linux/arm64
@@ -221,7 +221,7 @@ jobs:
 
             - uses: docker/setup-buildx-action@v4
 
-            - uses: docker/build-push-action@v7.1.0
+            - uses: docker/build-push-action@v7.2.0
               with:
                   context: .
                   platforms: ${{ matrix.platform }}
@@ -241,7 +241,7 @@ jobs:
                       PYTHON_VERSION=${{ matrix.python_version }}
                   labels: "org.opencontainers.image.source=${{ github.repositoryUrl }}"
 
-            - uses: docker/build-push-action@v7.1.0
+            - uses: docker/build-push-action@v7.2.0
               with:
                   context: .
                   platforms: ${{ matrix.platform }}
