+++ .pre-commit-config.yaml
@@ -15,7 +15,7 @@ repos:
       types_or: [c++, c, cuda]
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.15.13
+    rev: v0.15.14
     hooks:
     - id: ruff-check
       args: [--fix, --config=pyproject.toml]
