### 2026-05-29 b202ff4
build(deps): bump pyelftools from 0.32 to 0.33
Bumps [pyelftools](https://github.com/eliben/pyelftools) from 0.32 to 0.33.
- [Changelog](https://github.com/eliben/pyelftools/blob/main/CHANGES)
- [Commits](https://github.com/eliben/pyelftools/compare/v0.32...v0.33)

---
updated-dependencies:
- dependency-name: pyelftools
  dependency-version: '0.33'
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
