+++ pyproject.toml
@@ -13,7 +13,7 @@ requires = [
     "ijson",
     "mypy_extensions",
     "pandas",
-    "pyelftools==0.32",
+    "pyelftools==0.33",
     "pyyaml",
     "regex",
     "rich",
