### 2026-05-30 5fe9905
tests(tools): add a simple test for `get_symbol_table`
Signed-off-by: romintomasetti <romin.tomasetti@gmail.com>
