+++ docs/source/tests/tools/binaries/elf.rst
@@ -2,3 +2,4 @@ ELF
 ===
 
 .. automodule:: tests.tools.binaries.test_elf
+.. automodule:: tests.tools.binaries.test_symtab
