### 2026-06-05 442c5bf
build(deps): bump https://github.com/crate-ci/typos
Bumps [https://github.com/crate-ci/typos](https://github.com/crate-ci/typos) from v1.47.0 to 1.47.2.
- [Release notes](https://github.com/crate-ci/typos/releases)
- [Changelog](https://github.com/crate-ci/typos/blob/master/CHANGELOG.md)
- [Commits](https://github.com/crate-ci/typos/compare/v1.47.0...v1.47.2)

---
updated-dependencies:
- dependency-name: https://github.com/crate-ci/typos
  dependency-version: 1.47.2
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
