### 2026-06-20 2f4afbb
mypy fixes
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
