### 2026-06-27 1cfde7c
Fix for cuda core objectcode no longer being available in the objects.inv
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
