### 2026-07-03 621dee2
build(deps): bump https://github.com/crate-ci/typos
Bumps [https://github.com/crate-ci/typos](https://github.com/crate-ci/typos) from v1.47.2 to 1.48.0.
- [Release notes](https://github.com/crate-ci/typos/releases)
- [Changelog](https://github.com/crate-ci/typos/blob/master/CHANGELOG.md)
- [Commits](https://github.com/crate-ci/typos/compare/v1.47.2...v1.48.0)

---
updated-dependencies:
- dependency-name: https://github.com/crate-ci/typos
  dependency-version: 1.48.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
