### 2026-07-17 d0c90df
build(deps): bump https://github.com/pre-commit/mirrors-clang-format
Bumps [https://github.com/pre-commit/mirrors-clang-format](https://github.com/pre-commit/mirrors-clang-format) from v22.1.5 to 22.1.8.
- [Commits](https://github.com/pre-commit/mirrors-clang-format/compare/v22.1.5...v22.1.8)

---
updated-dependencies:
- dependency-name: https://github.com/pre-commit/mirrors-clang-format
  dependency-version: 22.1.8
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
