+++ .pre-commit-config.yaml
@@ -9,7 +9,7 @@ repos:
       exclude: '.*\.sass$'
 
   - repo: https://github.com/pre-commit/mirrors-clang-format
-    rev: v22.1.5
+    rev: v22.1.8
     hooks:
     - id: clang-format
       types_or: [c++, c, cuda]
