+++ requirements/requirements.python.txt
@@ -4,7 +4,7 @@ mypy==2.2.0
 myst-nb>=1.4.0
 nvtx>=0.2.15
 pylint>=4.0.6
-ruff==0.15.21
+ruff==0.15.22
 setuptools==83.0.0
 sphinx<=8.2.3
 sphinx-copybutton==0.5.2
