### 2026-07-20 77e8bc1
Split nsys in multiple files into subpackage
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
