### 2026-07-21 385bddc
Fixes for ruff update
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
