### 2026-07-21 356660a
Move documentation job to gpu image and runner
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
