+++ reprospect/testing/sass/matchers/add_int128.py
@@ -1,0 +1,238 @@
+import sys
+import typing
+
+from reprospect.testing.sass.composite import instructions_contain
+from reprospect.testing.sass.composite_impl import SequenceMatcher
+from reprospect.testing.sass.instruction import (
+    InstructionMatch,
+    OpcodeModsWithOperandsMatcher,
+    PatternBuilder,
+    RegisterMatcher,
+)
+from reprospect.testing.sass.instruction.constant import Constant
+from reprospect.testing.sass.instruction.instruction import ZeroOrOne
+from reprospect.testing.sass.instruction.register import Register
+from reprospect.tools.sass.decode import Instruction
+
+if sys.version_info >= (3, 12):
+    from typing import override
+else:
+    from typing_extensions import override
+
+class AddInt128Matcher(SequenceMatcher):
+    """
+    Match SASS instructions originating from the addition of 2 :code:`__int128`.
+
+    It may use one of several instruction patterns:
+
+    * :py:meth:`pattern_3IADD`
+    * :py:meth:`pattern_4IADD3`
+
+    .. note::
+
+        For a simple "load, add, store" sequence, the PTX may look as follows::
+
+            ld.global.nc.v2.u64 {%rd7, %rd8}, [%rd6];
+            add.s64 %rd9, %rd3, %rd5;
+            ld.global.v2.u64 {%rd10, %rd11}, [%rd9];
+            add.cc.s64 %rd12, %rd10, %rd7;
+            addc.cc.s64 %rd13, %rd11, %rd8;
+            st.global.v2.u64 [%rd9], {%rd12, %rd13};
+
+        According to https://docs.nvidia.com/cuda/parallel-thread-execution/#extended-precision-arithmetic-instructions-add-cc, the carry-out
+        value is written in the condition code register, usually ``P0``.
+
+    .. note::
+
+        It is not decorated with :py:func:`dataclasses.dataclass` because of https://github.com/mypyc/mypyc/issues/1061.
+    """
+    __slots__ = ('_index', 'start')
+
+    def __init__(self, start: str | None = None) -> None:
+        self._index: int = 0
+        self.start: typing.Final[str] = PatternBuilder.group(
+            start or Register.mod(PatternBuilder.any(Register.REG, Register.UREG), reuse=None, captured=False), group='start',
+        )
+
+    def pattern_3IADD(self, instructions: typing.Sequence[Instruction | str]) -> list[InstructionMatch] | None: # pylint: disable=invalid-name
+        """
+        Typically::
+
+            IADD.64 RZ, P0, R4, UR12
+            IADD.64 R4, R4, UR12
+            IADD.64.X R6, R6, UR14, P0
+        """
+        matched: list[InstructionMatch] = []
+
+        offset = 0
+
+        matcher_iadd_step_0 = OpcodeModsWithOperandsMatcher(
+            opcode='IADD', modifiers=('64',),
+            operands=(
+                'RZ',
+                Register.PRED,
+                self.start,
+                Register.mod(PatternBuilder.any(Register.REG, Register.UREG), reuse=None, captured=False),
+            ),
+        )
+        if (matched_iadd_step_0 := matcher_iadd_step_0.match(instructions[offset])) is None:
+            return None
+        offset += 1
+        matched.append(matched_iadd_step_0)
+
+        assert (iadd_step_0_reg_0 := RegisterMatcher(special=False).match(matched_iadd_step_0.operands[-2])) is not None
+        assert (iadd_step_0_reg_1 := RegisterMatcher(special=False).match(matched_iadd_step_0.operands[-1])) is not None
+
[…]
