+++ tests/test/sass/instruction/test_half.py
@@ -1,269 +1,0 @@
-import itertools
-import logging
-import pathlib
-import subprocess
-import typing
-
-import pytest
-
-from reprospect.test.features import PTX
-from reprospect.test.sass.composite import (
-    any_of,
-    findall,
-    findunique,
-    instruction_is,
-    instructions_contain,
-)
-from reprospect.test.sass.instruction import OpcodeModsMatcher
-from reprospect.test.sass.instruction.half import (
-    Fp16AddMatcher,
-    Fp16FusedMulAddMatcher,
-    Fp16MinMaxMatcher,
-    Fp16MulMatcher,
-)
-from reprospect.tools.binaries import CuObjDump
-from reprospect.tools.sass import Decoder
-
-from tests.parameters import PARAMETERS, Parameters
-from tests.test.sass.test_instruction import get_decoder
-
-
-class TestFp16FusedMulAddMatcher:
-    """
-    Tests for :py:class:`reprospect.test.sass.instruction.half.Fp16FusedMulAddMatcher`.
-    """
-    PTX: typing.Final[pathlib.Path] = pathlib.Path(__file__).parent / 'assets' / 'hfma2.ptx'
-
-    def test_individual(self) -> None:
-        matcher = Fp16FusedMulAddMatcher(packed=False)
-
-        assert (matched := matcher.match('HFMA2 R22, R22.H0_H0, R24.H0_H0, R27.H0_H0')) is not None
-        assert matched.operands == ('R22', 'R22.H0_H0', 'R24.H0_H0', 'R27.H0_H0')
-
-        assert (matched := matcher.match('HFMA2 R0, R0.H0_H0, R2.H0_H0, 3, 3')) is not None
-        assert matched.operands == ('R0', 'R0.H0_H0', 'R2.H0_H0', '3', '3')
-
-    def test_packed(self) -> None:
-        matcher = Fp16FusedMulAddMatcher(packed=True)
-
-        assert (matched := matcher.match(inst='HFMA2 R7, R2, R2, -RZ')) is not None
-        assert matched.additional is not None
-        assert matched.additional['dst'][0] == 'R7'
-        assert matched.operands == ('R7', 'R2', 'R2', '-RZ')
-
-        assert matcher.match('HFMA2 R19, -RZ, RZ, 0, 0') is not None
-        assert matcher.match('HFMA2.MMA R57, -RZ, RZ, 0, 0') is not None
-        assert matcher.match('HFMA2.MMA R20, -RZ, RZ, 1.875, 0') is not None
-        assert matcher.match('HFMA2.MMA R7, R0, 1, 1, R5') is not None
-
-        assert (matched := matcher.match('HFMA2 R24, -RZ, RZ, 0, 1.370906829833984375e-06')) is not None
-        assert matched.operands == ('R24', '-RZ', 'RZ', '0', '1.370906829833984375e-06')
-
-    def test_any(self) -> None:
-        matcher = Fp16FusedMulAddMatcher(packed=None)
-
-        assert matcher.match('HFMA2 R22, R22.H0_H0, R24.H0_H0, R27.H0_H0') is not None
-        assert matcher.match('HFMA2 R19, -RZ, RZ, 0, 0') is not None
-
-    @pytest.mark.parametrize('parameters', PARAMETERS, ids=str, scope='class')
-    def test_from_ptx(self, request, workdir: pathlib.Path, parameters: Parameters) -> None:
-        """
-        Compile the PTX from :py:attr:`PTX`.
-        """
-        cubin = workdir / f'{request.node.originalname}.{parameters.arch.as_sm}.cubin'
-        ptx   = workdir / f'{request.node.originalname}.{parameters.arch.as_sm}.ptx'
-
-        ptx_isa_version = PTX(arch=parameters.arch).min_isa_version
[…]
