### 2026-07-23 88e378d
Move installers to utils
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
