### 2026-07-23 cd2c86b
Move sass to binaries
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
