+++ reprospect/testing/sass/matchers/move32.py
@@ -9,7 +9,7 @@ from reprospect.testing.sass.instruction import (
 )
 from reprospect.testing.sass.instruction.operand import Operand
 from reprospect.testing.sass.instruction.register import Register
-from reprospect.tools.sass.decode import Instruction
+from reprospect.tools.binaries.sass.decode import Instruction
 
 if sys.version_info >= (3, 12):
     from typing import override
