+++ setup.py
@@ -57,8 +57,8 @@ def enable_mypyc(dist: Distribution) -> None:
             'reprospect/tools/nsys/cacher.py',
             'reprospect/tools/nsys/report.py',
             'reprospect/tools/nsys/session.py',
-            'reprospect/tools/sass/controlflow.py',
-            'reprospect/tools/sass/decode.py',
+            'reprospect/tools/binaries/sass/controlflow.py',
+            'reprospect/tools/binaries/sass/decode.py',
             'reprospect/utils/cmake.py',
             'reprospect/utils/compile_command.py',
             'reprospect/utils/detect.py',
