+++ examples/kokkos/atomic/example_add_int128.py
@@ -8,7 +8,7 @@ from reprospect.testing.sass.instruction import (
 )
 from reprospect.testing.sass.matchers import add_int128
 from reprospect.testing.sass.matchers.cas import AtomicCASMatcher
-from reprospect.tools.sass import ControlFlow, Decoder
+from reprospect.tools.binaries.sass import ControlFlow, Decoder
 
 from examples.kokkos.atomic import add, desul
 
