+++ reprospect/testing/sass/matchers/convert_fp_to_int.py
@@ -7,7 +7,7 @@ from reprospect.testing.sass.instruction import (
     OpcodeModsWithOperandsMatcher,
 )
 from reprospect.testing.sass.instruction.register import Register
-from reprospect.tools.sass.decode import Instruction
+from reprospect.tools.binaries.sass.decode import Instruction
 from reprospect.utils.types import ConvertibleTypeInfo, TypeInfo
 
 if sys.version_info >= (3, 12):
