### 2026-07-23 d6461ab
Move sass to binaries in testing too
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
