+++ tests/testing/sass/matchers/test_add_int128.py
@@ -1,98 +1,0 @@
-import logging
-import pathlib
-import re
-import subprocess
-import typing
-
-import pytest
-
-from reprospect.testing.sass.composite import instruction_is, instructions_contain
-from reprospect.testing.sass.instruction import LoadGlobalMatcher, RegisterMatcher
-from reprospect.testing.sass.matchers import add_int128
-from reprospect.utils import cmake
-
-from tests.parameters import PARAMETERS, Parameters
-from tests.testing.sass.test_instruction import get_decoder
-
-
-class TestAddInt128Matcher:
-    """
-    Tests for :py:class:`reprospect.testing.sass.matchers.add_int128.AddInt128Matcher`.
-    """
-    CODE_ADD_INT128: typing.Final[str] = """\
-__global__ void add_int128(__int128_t* __restrict__ const dst, const __int128_t* __restrict__ const src)
-{
-    const auto index = blockIdx.x * blockDim.x + threadIdx.x;
-    dst[index] += src[index];
-}
-"""
-
-    EXPECTED_PTX: typing.Final[re.Pattern[str]] = re.compile(r"""\
-add\.cc\.s64 %rd\d+, %rd\d+, %rd\d+;
-addc\.cc\.s64 %rd\d+, %rd\d+, %rd\d+;
-st\.global\.v2\.(u|b)64 \[%rd\d+\], {%rd\d+, %rd\d+};
-""")
-
-    def test_IADD(self) -> None:
-        INSTRUCTIONS: tuple[str, ...] = (
-            'IADD.64 RZ, P0, R4.reuse, R12.reuse',
-            'IADD.64 R4, R4, R12',
-            'IADD.64.X R6, R6, R14, P0',
-        )
-        matched = add_int128.AddInt128Matcher(start='R4.reuse').assert_matches(instructions=INSTRUCTIONS)
-
-        assert len(matched) == 3
-
-        assert matched[2].modifiers == ('64', 'X')
-        assert matched[2].operands == ('R6', 'R6', 'R14', 'P0')
-
-    def test_IADD3(self) -> None:
-        INSTRUCTIONS: tuple[str, ...] = (
-            'IADD3 R20, P0, PT, R4, R12, RZ',
-            'IADD3.X R21, P0, PT, R5, R13, RZ, P0, !PT',
-            'IADD3.X R22, P0, PT, R6, R14, RZ, P0, !PT',
-            'IADD3.X R23, PT, PT, R7, R15, RZ, P0, !PT',
-        )
-        matched = add_int128.AddInt128Matcher(start='R4').assert_matches(instructions=INSTRUCTIONS)
-
-        assert len(matched) == 4
-
-    @pytest.mark.parametrize('parameters', PARAMETERS, ids=str)
-    def test(self, request, parameters: Parameters, workdir: pathlib.Path, cmake_file_api: cmake.FileAPI) -> None:
-        FILE = workdir / f'{request.node.originalname}.{parameters.arch.as_sm}.cu'
-        FILE.write_text(self.CODE_ADD_INT128)
-
-        decoder, output = get_decoder(cwd=workdir, arch=parameters.arch, file=FILE, cmake_file_api=cmake_file_api, ptx=True)
-
-        ptx = subprocess.check_output(('cuobjdump', '--dump-ptx', output)).decode()
-
-        assert self.EXPECTED_PTX.search(ptx) is not None
-
-        # Find the global load of the source.
-        matcher_load_src = instructions_contain(matcher=LoadGlobalMatcher(arch=parameters.arch, size=128, readonly=True))
-        [matched_load_src] = matcher_load_src.assert_matches(instructions=decoder.instructions)
-
-        logging.info(matched_load_src)
-
-        # Find the global load of the destination.
[…]
