### 2026-07-24 66c8509
Move test ncu and test nsys to dedicated directories
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
