### 2026-07-24 56ba640
Move composite to sequence and the validators to validate
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
