+++ reprospect/testing/binaries/sass/operation/move32.py
@@ -5,10 +5,10 @@ from reprospect.testing.binaries.sass.instruction import (
     InstructionMatch,
     InstructionMatcher,
     OpcodeModsWithOperandsMatcher,
+    Operand,
     PatternBuilder,
+    Register,
 )
-from reprospect.testing.binaries.sass.instruction.operand import Operand
-from reprospect.testing.binaries.sass.instruction.register import Register
 from reprospect.tools.binaries.sass.decode import Instruction
 
 if sys.version_info >= (3, 12):
