+++ reprospect/tools/binaries/sass/controlflow.py
@@ -4,8 +4,10 @@ import typing
 
 import yaml
 
-from reprospect.testing.binaries.sass.instruction.instruction import Predicate
-from reprospect.testing.binaries.sass.instruction.pattern import PatternBuilder
+from reprospect.testing.binaries.sass.instruction import (
+    PatternBuilder,
+    Predicate,
+)
 from reprospect.tools.binaries.sass.decode import Instruction
 
 
