+++ reprospect/testing/binaries/sass/composite.py
@@ -1,291 +1,0 @@
-"""
-Thin user-facing factories for the matchers in :py:mod:`reprospect.testing.binaries.sass.composite_impl`.
-"""
-from __future__ import annotations
-
-import sys
-import typing
-
-from reprospect.testing.binaries.sass import composite_impl, instruction
-from reprospect.testing.binaries.sass.composite_impl import OperandMatcher
-from reprospect.tools.binaries.sass import Instruction
-
-if sys.version_info >= (3, 12):
-    from typing import override
-else:
-    from typing_extensions import override
-
-class Fluentizer(instruction.InstructionMatcher):
-    """
-    .. note::
-
-        It is not decorated with :py:func:`dataclasses.dataclass` because of https://github.com/mypyc/mypyc/issues/1061.
-    """
-    __slots__ = ('matcher',)
-
-    def __init__(self, matcher: instruction.InstructionMatcher) -> None:
-        self.matcher: typing.Final[instruction.InstructionMatcher] = matcher
-
-    def times(self, num: int) -> composite_impl.InSequenceAtMatcher | composite_impl.OrderedInSequenceMatcher:
-        """
-        Match :py:attr:`matcher` `num` times (consecutively).
-
-        .. note::
-
-            If `num` is 0, it tests for the absence of a match.
-        """
-        if num < 0:
-            raise RuntimeError
-        if num == 0:
-            raise NotImplementedError
-        if num == 1:
-            return composite_impl.InSequenceAtMatcher(matcher=self.matcher)
-        return composite_impl.OrderedInSequenceMatcher(matchers=(self.matcher,) * num)
-
-    def one_or_more_times(self) -> composite_impl.OneOrMoreInSequenceMatcher:
-        """
-        Match :py:attr:`matcher` one or more times (consecutively).
-        """
-        return composite_impl.OneOrMoreInSequenceMatcher(matcher=self.matcher)
-
-    def zero_or_more_times(self) -> composite_impl.ZeroOrMoreInSequenceMatcher:
-        """
-        Match :py:attr:`matcher` zero or more times (consecutively).
-        """
-        return composite_impl.ZeroOrMoreInSequenceMatcher(matcher=self.matcher)
-
-    def with_modifier(self, modifier: str, index: int | None = None) -> Fluentizer:
-        """
-        >>> from reprospect.testing.binaries.sass.composite import instruction_is
-        >>> from reprospect.testing.binaries.sass.instruction import AnyMatcher
-        >>> matcher = instruction_is(AnyMatcher()).with_modifier(modifier='U32').with_operand(index=0, operand='R4')
-        >>> matcher.match(inst='IMAD.WIDE.U32 R4, R7, 0x4, R4')
-        InstructionMatch(opcode='IMAD', modifiers=('WIDE', 'U32'), operands=('R4', 'R7', '0x4', 'R4'), predicate=None, additional=None)
-        """
-        return Fluentizer(matcher=composite_impl.ModifierValidator(
-            matcher=self.matcher,
-            index=index, modifier=modifier,
-        ))
-
-    def with_operand(self, operand: OperandMatcher, index: int | None = None) -> Fluentizer:
-        """
-        >>> from reprospect.testing.binaries.sass.composite   import instruction_is
-        >>> from reprospect.testing.binaries.sass.instruction import Fp32AddMatcher, RegisterMatcher
-        >>> from reprospect.tools.binaries.sass.decode     import RegisterType
[…]
