+++ examples/cuda/virtual_functions/example_dispatch.py
@@ -64,7 +64,7 @@ from reprospect.tools.binaries import (
 )
 from reprospect.tools.binaries.sass import ControlFlow, Decoder
 from reprospect.tools.binaries.sass.controlflow import BasicBlock, Graph
-from reprospect.tools.binaries.sass.decode import RegisterType
+from reprospect.tools.binaries.sass.decoder import RegisterType
 from reprospect.utils import detect
 
 if sys.version_info >= (3, 11):
