+++ reprospect/tools/binaries/sass/__init__.py
@@ -1,5 +1,5 @@
 from .controlflow import ControlFlow
-from .decode import ControlCode, Decoder, Instruction
+from .decoder import ControlCode, Decoder, Instruction
 
 __all__ = (
     'ControlCode',
