### 2026-07-24 7c5b56d
Move tests in binaries sass
