+++ docs/source/tests/testing/binaries/sass/instruction.rst
@@ -1,12 +1,14 @@
 Instruction
 ===========
 
+.. automodule:: tests.testing.binaries.sass.instruction.test_any
+.. automodule:: tests.testing.binaries.sass.instruction.test_atomic
+.. automodule:: tests.testing.binaries.sass.instruction.test_branch
 .. automodule:: tests.testing.binaries.sass.instruction.test_address
 .. automodule:: tests.testing.binaries.sass.instruction.test_constant
 .. automodule:: tests.testing.binaries.sass.instruction.test_cublas
 .. automodule:: tests.testing.binaries.sass.instruction.test_fp32add
 .. automodule:: tests.testing.binaries.sass.instruction.test_fp64add
-.. automodule:: tests.testing.binaries.sass.instruction.test_half
 .. automodule:: tests.testing.binaries.sass.instruction.test_immediate
 .. automodule:: tests.testing.binaries.sass.instruction.test_integer
 .. automodule:: tests.testing.binaries.sass.instruction.test_operand
