### 2026-07-24 ebcbcfd
Move environment to utils and compile command to tools
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
