+++ setup.py
@@ -46,10 +46,11 @@ def enable_mypyc(dist: Distribution) -> None:
             'reprospect/testing/binaries/sass/sequence/sequence.py',
             'reprospect/tools/architecture.py',
             'reprospect/tools/binaries/cuobjdump.py',
-            'reprospect/tools/binaries/demangle.py',
+            'reprospect/tools/binaries/demangler.py',
             'reprospect/tools/binaries/elf.py',
             'reprospect/tools/binaries/nvdisasm.py',
             'reprospect/tools/binaries/symtab.py',
+            'reprospect/tools/compile_command.py',
             'reprospect/tools/ncu/cacher.py',
             'reprospect/tools/ncu/metrics.py',
             'reprospect/tools/ncu/report.py',
@@ -60,7 +61,6 @@ def enable_mypyc(dist: Distribution) -> None:
             'reprospect/tools/binaries/sass/controlflow.py',
             'reprospect/tools/binaries/sass/decoder.py',
             'reprospect/utils/cmake.py',
-            'reprospect/utils/compile_command.py',
             'reprospect/utils/detect.py',
             'reprospect/utils/ldd.py',
             'reprospect/utils/nvcc.py',
