+++ docs/source/tests/tools.rst
@@ -6,7 +6,7 @@ Tests for :py:mod:`reprospect.tools`.
 The tests for the three main subpackages can be found in:
 
 .. toctree::
-   :maxdepth: 3
+   :maxdepth: 2
 
    tools/binaries
    tools/ncu
@@ -22,6 +22,11 @@ Cacher
 ------
 
 .. automodule:: tests.tools.test_cacher
+
+Compile command
+---------------
+
+.. automodule:: tests.tools.test_compile_command
 
 Device properties
 -----------------
