### 2026-07-24 63b0401
Expand notebook example on binary analysis
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
