+++ examples/kokkos/view/example_restrict.py
@@ -132,8 +132,8 @@ from reprospect.testing.binaries.sass.sequence import (
 )
 from reprospect.tools import ncu
 from reprospect.tools.binaries.cuobjdump import CuObjDump
-from reprospect.tools.binaries.sass.controlflow import ControlFlow, Graph
-from reprospect.tools.binaries.sass.decoder import Decoder
+from reprospect.tools.binaries.sass import ControlFlow, Decoder
+from reprospect.tools.binaries.sass.controlflow import Graph
 from reprospect.utils import detect, environment
 
 if sys.version_info >= (3, 12):
