+++ docs/source/getting-started.rst
@@ -4,4 +4,4 @@ Getting started
 .. toctree::
    :maxdepth: 3
 
-   getting-started/example_atomics.ipynb
+   getting-started/example_binary_analysis.ipynb
