+++ .github/workflows/build.yml
@@ -34,7 +34,7 @@ jobs:
       runs-on: ubuntu-latest
       steps:
         - uses: actions/checkout@v7
-        - uses: actions/setup-python@v6.2.0
+        - uses: actions/setup-python@v7
           with:
             python-version: '3.14'
         - uses: pre-commit/action@v3.0.1
@@ -327,7 +327,7 @@ jobs:
         if: ${{ ! failure() && ! cancelled() }}
         steps:
             - uses: actions/checkout@v7
-            - uses: actions/setup-python@v6
+            - uses: actions/setup-python@v7
               with:
                   python-version: '3.11'
             - run: python -m pip install -U cibuildwheel
@@ -348,7 +348,7 @@ jobs:
             image: ubuntu:24.04
         steps:
             - uses: actions/checkout@v7
-            - uses: actions/setup-python@v6
+            - uses: actions/setup-python@v7
               with:
                   python-version: '3.11'
             - run: python -m pip install -U build
