### 2026-07-24 e898183
build(deps): bump sphinxcontrib-mermaid from 2.0.3 to 2.1.0
Bumps [sphinxcontrib-mermaid](https://github.com/mgaitan/sphinxcontrib-mermaid) from 2.0.3 to 2.1.0.
- [Changelog](https://github.com/mgaitan/sphinxcontrib-mermaid/blob/master/CHANGELOG.md)
- [Commits](https://github.com/mgaitan/sphinxcontrib-mermaid/compare/2.0.3...2.1.0)

---
updated-dependencies:
- dependency-name: sphinxcontrib-mermaid
  dependency-version: 2.1.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
