### 2026-07-28 1e116b5
Newly encountered operand SR_GLOBALTIMERLO
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
