### 2026-07-28 8e8609b
Notebook api tracing
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
