+++ docs/source/getting-started.rst
@@ -4,4 +4,5 @@ Getting started
 .. toctree::
    :maxdepth: 3
 
+   getting-started/example_api_tracing.ipynb
    getting-started/example_binary_analysis.ipynb
