### 2026-08-03 8955f71
Add metric l1tex cache global store sectors
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
