+++ reprospect/tools/ncu/__init__.py
@@ -10,6 +10,7 @@ from .metrics import (
     L1TEXCacheGlobalLoadWavefronts,
     L1TEXCacheGlobalStore,
     L1TEXCacheGlobalStoreInstructions,
+    L1TEXCacheGlobalStoreRequests,
     L1TEXCacheGlobalStoreSectors,
     L1TEXCacheLocalStore,
     L1TEXCacheLocalStoreInstructions,
@@ -20,8 +21,10 @@ from .metrics import (
     MetricCounter,
     MetricCounterRollUp,
     MetricDeviceAttribute,
+    MetricKind,
     MetricRatio,
     MetricRatioRollUp,
+    gather,
 )
 from .report import (
     ProfilingMetrics,
@@ -43,6 +46,7 @@ __all__ = (
     'L1TEXCacheGlobalLoadWavefronts',
     'L1TEXCacheGlobalStore',
     'L1TEXCacheGlobalStoreInstructions',
+    'L1TEXCacheGlobalStoreRequests',
     'L1TEXCacheGlobalStoreSectors',
     'L1TEXCacheLocalStore',
     'L1TEXCacheLocalStoreInstructions',
@@ -53,10 +57,12 @@ __all__ = (
     'MetricCounter',
     'MetricCounterRollUp',
     'MetricDeviceAttribute',
+    'MetricKind',
     'MetricRatio',
     'MetricRatioRollUp',
     'ProfilingMetrics',
     'ProfilingResults',
     'Report',
     'Session',
+    'gather',
 )
