### 2026-08-03 071c655
Uniformise metric factories
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
