### 2026-08-03 5afd54b
Allow paths of submetric names. Add counter roll-up quantities.
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
