+++ reprospect/tools/ncu/__init__.py
@@ -20,6 +20,7 @@ from .metrics import (
     MetricCorrelation,
     MetricCounter,
     MetricCounterRollUp,
+    MetricCounterRollUpQuantity,
     MetricDeviceAttribute,
     MetricKind,
     MetricRatio,
@@ -56,6 +57,7 @@ __all__ = (
     'MetricCorrelation',
     'MetricCounter',
     'MetricCounterRollUp',
+    'MetricCounterRollUpQuantity',
     'MetricDeviceAttribute',
     'MetricKind',
     'MetricRatio',
