### 2026-08-03 c7f775e
Add labels to metric
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
