+++ tests/tools/ncu/test_ncu.py
@@ -28,7 +28,7 @@ class TestProfilingResults:
             data=ncu.ProfilingMetrics({
                 'smsp__inst_executed.sum': 100.,
                 'sass__inst_executed_per_opcode': MetricCorrelationData(correlated={'LDCU': 16., 'LDC': 16.}),
-                'L1/TEX cache global load sectors.sum': 0.,
+                'L1/TEX cache global load sectors sum': 0.,
             }),
         )
         results.assign_metrics(
@@ -36,7 +36,7 @@ class TestProfilingResults:
             data=ncu.ProfilingMetrics({
                 'smsp__inst_executed.sum': 96.,
                 'sass__inst_executed_per_opcode': MetricCorrelationData(correlated={'LDCU': 16., 'LDC': 16.}),
-                'L1/TEX cache global load sectors.sum': 0.,
+                'L1/TEX cache global load sectors sum': 0.,
             }),
         )
         return results
@@ -153,13 +153,13 @@ Profiling results
     │       └── kernel
     │           ├── smsp__inst_executed.sum: 100.0
     │           ├── sass__inst_executed_per_opcode: MetricCorrelationData(correlated={'LDCU': 16.0, 'LDC': 16.0}, value=None)
-    │           └── L1/TEX cache global load sectors.sum: 0.0
+    │           └── L1/TEX cache global load sectors sum: 0.0
     └── nvtx_push_region_B
         └── nvtx_push_region_other_kernel
             └── other_kernel
                 ├── smsp__inst_executed.sum: 96.0
                 ├── sass__inst_executed_per_opcode: MetricCorrelationData(correlated={'LDCU': 16.0, 'LDC': 16.0}, value=None)
-                └── L1/TEX cache global load sectors.sum: 0.0
+                └── L1/TEX cache global load sectors sum: 0.0
 """
 
         results_A = results.query(("nvtx_range_name", "nvtx_push_region_A"))
@@ -170,7 +170,7 @@ Profiling results
     └── kernel
         ├── smsp__inst_executed.sum: 100.0
         ├── sass__inst_executed_per_opcode: MetricCorrelationData(correlated={'LDCU': 16.0, 'LDC': 16.0}, value=None)
-        └── L1/TEX cache global load sectors.sum: 0.0
+        └── L1/TEX cache global load sectors sum: 0.0
 """
 
 class TestCommand:
@@ -203,29 +203,130 @@ class TestMetrics:
     """
     Tests for :py:mod:`reprospect.tools.ncu.metrics`.
     """
-    METRICS: typing.Final[tuple[tuple[ncu.MetricKind, tuple[str, ...]], ...]] = (
-        (ncu.MetricCounter(name='dram__bytes_op_write', subs=((ncu.MetricCounterRollUp.SUM, ncu.MetricCounterRollUpQuantity.PCT_OF_PEAK_SUSTAINED_ELAPSED),)), ('dram__bytes_op_write.sum.pct_of_peak_sustained_elapsed',)),
+    class MetricCase(typing.NamedTuple):
+        metric: ncu.MetricKind
+        gathered: tuple[str, ...]
+        labels: tuple[str, ...]
+
+    class CreatedMetricsCase(typing.NamedTuple):
+        metrics: tuple[ncu.MetricKind, ...]
+        expected: tuple[ncu.MetricKind, ...]
+        gathered: tuple[str, ...]
+        labels: tuple[str, ...]
+
+    METRICS: typing.Final[tuple[MetricCase, ...]] = (
+        MetricCase(
[…]
