### 2026-08-03 c37fcbe
Add warp stall metric factories
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
