### 2026-08-04 c15aa27
Do not display default sub labels
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
