+++ reprospect/tools/ncu/metrics.py
@@ -10,13 +10,13 @@ if sys.version_info >= (3, 11):
 else:
     from backports.strenum.strenum import StrEnum
 
+
 #: A single metric value type.
 ValueType: typing.TypeAlias = int | float
 
 #: A single metric value type or a dictionary of submetric values of such type.
 MetricData: typing.TypeAlias = ValueType | dict[str, ValueType]
 
-@dataclasses.dataclass(frozen=True, slots=True)
 class Metric:
     """
     Used to represent a ``ncu`` metric.
@@ -27,20 +27,30 @@ class Metric:
     References:
 
     * https://docs.nvidia.com/nsight-compute/ProfilingGuide/index.html#metrics-structure
+
+    .. note::
+
+        It is not decorated with :py:func:`dataclasses.dataclass` because of https://github.com/mypyc/mypyc/issues/1061.
     """
-    #: The base name of the metric.
-    name: str
+    #: Sub-metric components implied by the metric kind, omitted from labels.
+    SILENT_SUBS: typing.ClassVar[frozenset[str]] = frozenset()
 
-    #: Human readable name.
-    pretty_name: str | None = None
+    __slots__ = ('name', 'pretty_name', 'subs')
 
-    #: Optional sub-metric names, each one stored as a path of components
-    subs: tuple[str | tuple[str, ...], ...] | None = None
+    def __init__(self, name: str, pretty_name: str | None = None, subs: tuple[str | tuple[str, ...], ...] | None = None) -> None:
+        """
+        :param name: The base name of the metric.
+        :param pretty_name: Human readable name; defaults to :py:attr:`name`.
+        :param subs: Sub-metric names. The constructor normalizes each to a path of components.
+        """
+        self.name: typing.Final[str] = name
+        self.pretty_name: typing.Final[str | None] = pretty_name
+        self.subs: typing.Final[tuple[tuple[str, ...], ...] | None] = tuple((sub,) if isinstance(sub, str) else sub for sub in subs) if subs is not None else None
 
-    def __post_init__(self) -> None:
-        if self.subs is not None:
-            object.__setattr__(self, 'subs',
-                tuple((sub,) if isinstance(sub, str) else sub for sub in self.subs))
+    def __eq__(self, other: object) -> bool:
+        if isinstance(other, Metric):
+            return self.name == other.name and self.pretty_name == other.pretty_name and self.subs == other.subs
+        return NotImplemented
 
     def gather(self) -> tuple[str, ...]:
         """
@@ -54,16 +64,25 @@ class Metric:
         """
         Get the list of sub-metric labels. Parallel to :py:meth:`gather`, but uses the pretty name.
         """
-        if self.pretty_name is not None:
-            if self.subs is not None:
-                return tuple(' '.join((self.pretty_name, *sub)) for sub in self.subs)
+        if self.pretty_name is None:
+            return self.gather()
+        if self.subs is None:
             return (self.pretty_name,)
-        return self.gather()
+        result = []
+        for sub in self.subs:
+            sub_labels = tuple(getattr(c, 'label', c) for c in sub if c not in self.SILENT_SUBS)
[…]
