### 2026-08-04 9643215
Further label improvements
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
