### 2026-08-05 49e7da4
Replace our use of rich tools df_to_table with our own. Add rows_to_table.
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
