+++ pyproject.toml
@@ -17,7 +17,6 @@ requires = [
     "pyyaml",
     "regex",
     "rich",
-    "rich-tools",
     "semantic_version",
     "system-helpers>=0.0.3",
     "typing-extensions>=4.4.0",
