### 2026-08-05 232145c
Add kernel profiling notebook
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
