### 2026-08-07 ee04431
build(deps): bump coverage from 7.15.2 to 7.15.3
Bumps [coverage](https://github.com/coveragepy/coveragepy) from 7.15.2 to 7.15.3.
- [Release notes](https://github.com/coveragepy/coveragepy/releases)
- [Changelog](https://github.com/coveragepy/coveragepy/blob/main/CHANGES.rst)
- [Commits](https://github.com/coveragepy/coveragepy/compare/7.15.2...7.15.3)

---
updated-dependencies:
- dependency-name: coverage
  dependency-version: 7.15.3
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
