+++ requirements/requirements.python.txt
@@ -1,4 +1,4 @@
-coverage==7.15.2
+coverage==7.15.3
 matplotlib
 mypy==2.3.0
 myst-nb>=1.4.0
