+++ .github/workflows/build.yml
@@ -117,7 +117,7 @@ jobs:
             - uses: actions/checkout@v7
             - run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
 
-            - uses: docker/login-action@v4.5.2
+            - uses: docker/login-action@v4.6.0
               with:
                   registry: ghcr.io
                   username: ${{ github.actor }}
@@ -170,7 +170,7 @@ jobs:
         steps:
             - uses: actions/checkout@v7
 
-            - uses: docker/login-action@v4.5.2
+            - uses: docker/login-action@v4.6.0
               with:
                   registry: ghcr.io
                   username: ${{ github.actor }}
@@ -213,7 +213,7 @@ jobs:
             - uses: actions/checkout@v7
             - run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
 
-            - uses: docker/login-action@v4.5.2
+            - uses: docker/login-action@v4.6.0
               with:
                   registry: ${{ env.REGISTRY }}
                   username: ${{ github.actor }}
