### 2026-08-09 c3f535a
Fix pandas 2.3.3 vs 3.0.5 missing value behavior
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
