### 2026-08-10 573d575
Centralize common dependency information
