+++ .github/workflows/strategy.py
@@ -4,6 +4,7 @@ import argparse
 import copy
 import dataclasses
 import enum
+import functools
 import json
 import logging
 import pathlib
@@ -53,8 +54,6 @@ AVAILABLE_RUNNERS: typing.Final[tuple[Runner, ...]] = (
 
 SELF_HOSTED: typing.Final[tuple[str, ...]] = ('self-hosted', 'linux', 'docker')
 
-KOKKOS_SHA: typing.Final[str] = '5.1.0'
-
 @dataclasses.dataclass(frozen=False, slots=True)
 class Compiler:
     ID: COMPILER_ID
@@ -74,6 +73,8 @@ class Compiler:
                     raise ValueError
 
 JobDict: typing.TypeAlias = dict[str, typing.Any]
+
+DependencyDict: typing.TypeAlias = dict[str, dict[str, str]]
 
 @dataclasses.dataclass(frozen=False, slots=True)
 class Config:
@@ -222,7 +223,7 @@ def runs_on(spec: Runner, jtype: typing.Literal['tests', 'examples', 'documentat
         case _:
             raise ValueError
 
-def complete_job_impl(*, partial: JobDict, args: argparse.Namespace) -> JobDict:
+def complete_job_impl(*, partial: JobDict, deps: DependencyDict, args: argparse.Namespace) -> JobDict:
     """
     Add fields to a job.
     """
@@ -250,8 +251,8 @@ def complete_job_impl(*, partial: JobDict, args: argparse.Namespace) -> JobDict:
     partial['nvidia_arch'] = str(arch)
 
     partial['base_image'] = f'{base_name}:{base_tag}@{base_digest}'
-    partial[     'image'] = full_image(platform=partial['platform'], args=args, name=name,                          tag=base_tag)
-    partial[    'kokkos'] = full_image(platform=partial['platform'], args=args, name=f'{name}-kokkos-{KOKKOS_SHA}', tag=f'{base_tag}-{arch}'.lower())
+    partial[     'image'] = full_image(platform=partial['platform'], args=args, name=name,                                     tag=base_tag)
+    partial[    'kokkos'] = full_image(platform=partial['platform'], args=args, name=f'{name}-kokkos-{deps["kokkos"]["sha"]}', tag=f'{base_tag}-{arch}'.lower())
 
     # Name of the job.
     # See also https://futurestud.io/tutorials/github-actions-customize-the-job-name.
@@ -304,35 +305,31 @@ def complete_job_impl(*, partial: JobDict, args: argparse.Namespace) -> JobDict:
 
     return partial
 
-def from_config(config: Config, args: argparse.Namespace) -> list[JobDict]:
+def from_config(config: Config, deps: DependencyDict, args: argparse.Namespace) -> list[JobDict]:
     return [
-        complete_job_impl(partial=copy.deepcopy(job), args=args)
+        complete_job_impl(partial=copy.deepcopy(job), deps=deps, args=args)
         for job in config.jobs()
     ]
 
-def common_dependencies() -> dict:
+@functools.cache
+def dependencies() -> DependencyDict:
     """
     Read the information on dependencies common across all strategy entries.
     """
-    return json.loads((pathlib.Path(__file__).parents[2] / 'dependencies.json').read_text())
-
-def common_dependency_versions_as_build_args() -> str:
-    """
-    Return the common dependency versions as build arguments.
-    """
-    deps = common_dependencies()
-    return '\n'.join((
-        f'DOXYGEN_VERSION={deps["doxygen"]["version"]}',
-        f'GOOGLEBENCHMARK_VERSION={deps["googlebenchmark"]["version"]}',
[…]
