### 2026-08-10 3d90fd2
Print list of images in documentation
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
