### 2026-08-10 b11d91f
Add instructions for running examples and tests
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
