### 2026-08-12 226a00e
Add Kokkos CMake configuration details
Signed-off-by: Maarten Arnst <maarten.arnst@uliege.be>
