+++ requirements/requirements.python.txt
@@ -5,7 +5,7 @@ myst-nb>=1.4.0
 nvtx>=0.2.15
 pylint>=4.0.6
 ruff==0.16.1
-setuptools==83.0.0
+setuptools==84.0.0
 sphinx<=8.2.3
 sphinx-copybutton==0.5.2
 sphinx-rtd-theme>=3.1.0
